Lightning.Auditing.Audit behaviour (Lightning v0.12.1)

Macro module to add common model behaviour to a given Ecto model

Callbacks

update_changes(changes)

@callback update_changes(changes :: map()) :: map()

Functions

event(item_type, event, item_id, actor_id, changes \\ %{}, update_fun \\ fn x -> x end)

@spec event(
  String.t(),
  String.t(),
  Ecto.UUID.t(),
  Ecto.UUID.t(),
  Ecto.Changeset.t() | map() | nil,
  update_changes_fun :: (map() -> map())
) :: :no_changes | Ecto.Changeset.t()

Creates a changeset for the event identified by item_id and caused by actor_id.

The given changes can be either nil, Ecto.Changeset, struct or map.

It returns :no_changes in case of an Ecto.Changeset changes that changed nothing or an Ecto.Changeset with the event ready to be inserted.

save(changes, repo)

@spec save(Ecto.Changeset.t() | :no_changes, Ecto.Repo.t()) ::
  {:ok, :no_changes} | {:ok, Ecto.Schema.t()} | {:error, Ecto.Changeset.t()}

Saves the event to the Repo.

In case of nothing changes, do nothing.

It returns {:ok, :no_changes} if nothing changed, {:ok, struct} if the log has been successfully saved or {:error, changeset} in case of error.