Lightning.CLI (Lightning v2.15.11)

Module providing facilities to make calls to the OpenFn CLI.

See @openfn/cli

Functions

execute(command)

@spec execute(command :: String.t()) :: Lightning.CLI.Result.t()

Execute a command in a child process and parse the results.

metadata(state, adaptor_path)

@spec metadata(state :: map(), adaptor_path :: String.t()) :: Lightning.CLI.Result.t()

Retrieve metadata for a given adaptor and configuration.