Lightning.ChangesetUtils (Lightning v2.4.9)

Extra functions for Ecto.Changeset.

Functions

put_if_provided(changeset, field, params)

@spec put_if_provided(Ecto.Changeset.t(), atom(), map()) :: Ecto.Changeset.t()

Puts a new change in the changeset if the params have the field.

Useful when params don't have a given key and you don't want to set the field to nil.

put_new_change(changeset, field, value)

@spec put_new_change(Ecto.Changeset.t(), atom(), any()) :: Ecto.Changeset.t()

Puts a new change in the changeset if the field is not already set.

NOTE: This function considers a field with a nil value as not set.