Lightning.Collaborate (Lightning v2.15.0-pre4)

Public API for starting collaborative workflow editing sessions.

This module serves as the main entry point for collaborative editing, coordinating the creation and management of document and session processes for workflow collaboration.

All collaborative sessions require a workflow struct.

Example

# Existing workflow
workflow = Lightning.Workflows.get_workflow(workflow_id)

Collaborate.start(user: user, workflow: workflow)

# New workflow
workflow = %Lightning.Workflows.Workflow{
  id: workflow_id,
  project_id: project_id,
  name: "",
  positions: %{}
}

Collaborate.start(user: user, workflow: workflow)

Functions

lookup(key)

See Lightning.Collaboration.Registry.lookup/1.

start(opts)

@spec start(opts :: Keyword.t()) :: GenServer.on_start()

start_document(workflow, document_name)

whereis(key)

See Lightning.Collaboration.Registry.whereis/1.