Lightning.Config behaviour (Lightning v2.10.7)

Centralised runtime configuration for Lightning.

Callbacks

apollo(key)

@callback apollo(key :: atom() | nil) :: map()

check_flag?(atom)

@callback check_flag?(atom()) :: boolean() | nil

cors_origin()

@callback cors_origin() :: list()

default_max_run_duration()

@callback default_max_run_duration() :: integer()

email_sender_name()

@callback email_sender_name() :: String.t()

get_extension_mod(key)

@callback get_extension_mod(key :: atom()) :: any()

google(key)

@callback google(key :: atom()) :: any()

grace_period()

@callback grace_period() :: integer()

instance_admin_email()

@callback instance_admin_email() :: String.t()

kafka_alternate_storage_enabled?()

@callback kafka_alternate_storage_enabled?() :: boolean()

kafka_alternate_storage_file_path()

@callback kafka_alternate_storage_file_path() :: String.t()

kafka_duplicate_tracking_retention_seconds()

@callback kafka_duplicate_tracking_retention_seconds() :: integer()

kafka_notification_embargo_seconds()

@callback kafka_notification_embargo_seconds() :: integer()

kafka_number_of_consumers()

@callback kafka_number_of_consumers() :: integer()

kafka_number_of_messages_per_second()

@callback kafka_number_of_messages_per_second() :: float()

kafka_number_of_processors()

@callback kafka_number_of_processors() :: integer()

kafka_triggers_enabled?()

@callback kafka_triggers_enabled?() :: boolean()

oauth_provider(key)

@callback oauth_provider(key :: atom()) :: keyword() | nil

purge_deleted_after_days()

@callback purge_deleted_after_days() :: integer()

repo_connection_token_signer()

@callback repo_connection_token_signer() :: Joken.Signer.t()

reset_password_token_validity_in_days()

@callback reset_password_token_validity_in_days() :: integer()

run_token_signer()

@callback run_token_signer() :: Joken.Signer.t()

storage()

@callback storage() :: term()

storage(key)

@callback storage(key :: atom()) :: term()

token_signer()

@callback token_signer() :: Joken.Signer.t()

usage_tracking()

@callback usage_tracking() :: Keyword.t()

usage_tracking_cleartext_uuids_enabled?()

@callback usage_tracking_cleartext_uuids_enabled?() :: boolean()

usage_tracking_cron_opts()

@callback usage_tracking_cron_opts() :: [Oban.Plugins.Cron.cron_input()]

usage_tracking_enabled?()

@callback usage_tracking_enabled?() :: boolean()

usage_tracking_host()

@callback usage_tracking_host() :: String.t()

usage_tracking_run_chunk_size()

@callback usage_tracking_run_chunk_size() :: integer()

worker_secret()

@callback worker_secret() :: binary() | nil

worker_token_signer()

@callback worker_token_signer() :: Joken.Signer.t()

Functions

apollo(key \\ nil)

Returns the Apollo server configuration.

check_flag?(flag)

cors_origin()

default_max_run_duration()

Returns the default maximum run duration in seconds.

email_sender_name()

get_extension_mod(key)

google(key)

grace_period()

The grace period is configurable and is used to wait for an additional amount of time after a given run was meant to be finished.

The returned value is in seconds.

instance_admin_email()

kafka_alternate_storage_enabled?()

kafka_alternate_storage_file_path()

kafka_duplicate_tracking_retention_seconds()

kafka_notification_embargo_seconds()

kafka_number_of_consumers()

kafka_number_of_messages_per_second()

kafka_number_of_processors()

kafka_triggers_enabled?()

oauth_provider(key)

purge_deleted_after_days()

repo_connection_token_signer()

reset_password_token_validity_in_days()

run_token_signer()

Returns the Token signer used to sign and verify run tokens.

storage()

storage(key)

token_signer()

usage_tracking()

usage_tracking_cleartext_uuids_enabled?()

usage_tracking_cron_opts()

usage_tracking_enabled?()

usage_tracking_host()

usage_tracking_run_chunk_size()

worker_secret()

worker_token_signer()

Returns the Token signer used to verify worker tokens.