Lightning.Credentials (Lightning v0.1.5)

The Credentials context.

Functions

change_credential(credential, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking credential changes.

Examples

iex> change_credential(credential)
%Ecto.Changeset{data: %Credential{}}

create_credential(attrs \\ %{})

Creates a credential.

Examples

iex> create_credential(%{field: value})
{:ok, %Credential{}}

iex> create_credential(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_credential(credential)

Deletes a credential.

Examples

iex> delete_credential(credential)
{:ok, %Credential{}}

iex> delete_credential(credential)
{:error, %Ecto.Changeset{}}

get_credential!(id)

Gets a single credential.

Raises Ecto.NoResultsError if the Credential does not exist.

Examples

iex> get_credential!(123)
%Credential{}

iex> get_credential!(456)
** (Ecto.NoResultsError)

get_credential_body(id)

@spec get_credential_body(
  credential_or_uuid :: Lightning.Credentials.Credential.t() | Ecto.UUID.t()
) ::
  String.t() | nil

Gets a single credential body as a string Returns nil if the Credential does not exist.

list_credentials()

Returns the list of credentials.

Examples

iex> list_credentials()
[%Credential{}, ...]

list_credentials(project)

list_credentials_for_user(user_id)

Returns the list of credentials for a given user.

Examples

iex> list_credentials_for_user(123)
[%Credential{user_id: 123}, %Credential{user_id: 123},...]

update_credential(credential, attrs)

Updates a credential.

Examples

iex> update_credential(credential, %{field: new_value})
{:ok, %Credential{}}

iex> update_credential(credential, %{field: bad_value})
{:error, %Ecto.Changeset{}}