Lightning.Credentials.SensitiveValues (Lightning v0.4.5)

Functions to pull out sensitive values inside a credential.

These values are used to scrub logs for leaked secrets.

Types

pairs()

@type pairs() :: {String.t(), String.t() | number() | boolean()}

raw_pairs()

@type raw_pairs() :: {String.t(), pairs() | map()}

Functions

flatten_map(item)

@spec flatten_map(item :: %{required(String.t()) => any()} | [raw_pairs()]) :: [
  pairs()
]

Keys that are not considered sensitive

secret_values(map)

@spec secret_values(map :: %{required(String.t()) => any()}) :: [
  String.t() | number() | boolean()
]

Given a map, find all values allowed (via @safe_keys) and return them as a list.