Lightning.Helpers (Lightning v0.8.1)

Common functions for the context

Functions

coerce_json_field(attrs, field)

@spec coerce_json_field(map(), Map.key()) :: map()

Changes a given maps field from a json string to a map. If it cannot be converted, it leaves the original value

indefinite_article(noun)

json_safe(data)

Recursively ensures a given map is safe to convert to JSON, where all keys are strings and all values are json safe (primitive values).

ms_to_human(milliseconds)

@spec ms_to_human(integer()) :: String.t() | {:error, :invalid_duration}

Converts milliseconds (integer) to a human duration, such as "1 minute" or "45 years, 6 months, 5 days, 21 hours, 12 minutes, 34 seconds" using Timex.Format.Duration.Formatters.Humanized.format().