Lightning.Invocation (Lightning v0.2.0)

The Invocation context.

Functions

change_dataclip(dataclip, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking dataclip changes.

Examples

iex> change_dataclip(dataclip)
%Ecto.Changeset{data: %Dataclip{}}

change_run(run, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking run changes.

Examples

iex> change_run(run)
%Ecto.Changeset{data: %Run{}}

create(event_attrs)

create(event_attrs, dataclip_attrs)

@spec create(
  %{job_id: binary(), project_id: binary(), type: :webhook | :cron},
  %{
    body: map(),
    project_id: binary(),
    type: Lightning.Invocation.Dataclip.source_type()
  }
) ::
  {:ok | :error,
   %{
     event: Lightning.Invocation.Event,
     run: Lightning.Invocation.Run,
     dataclip: Lightning.Invocation.Dataclip
   }}

Create a new invocation based on a job and a body of data, which gets saved as a Dataclip; resulting in a Run associated with the Event.

create_dataclip(attrs \\ %{})

@spec create_dataclip(attrs :: map()) ::
  {:ok, Lightning.Invocation.Dataclip.t()}
  | {:error, Ecto.Changeset.t(Lightning.Invocation.Dataclip)}

Creates a dataclip.

Examples

iex> create_dataclip(%{field: value})
{:ok, %Dataclip{}}

iex> create_dataclip(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_event(attrs \\ %{})

Creates an event.

Examples

iex> create_event(%{type: :webhook, dataclip_id: dataclip.id})
{:ok, %Dataclip{}}

iex> create_dataclip(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_run(attrs \\ %{})

Creates a run.

Examples

iex> create_run(%{field: value})
{:ok, %Run{}}

iex> create_run(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_dataclip(dataclip)

Deletes a dataclip.

Examples

iex> delete_dataclip(dataclip)
{:ok, %Dataclip{}}

iex> delete_dataclip(dataclip)
{:error, %Ecto.Changeset{}}

delete_run(run)

Deletes a run.

Examples

iex> delete_run(run)
{:ok, %Run{}}

iex> delete_run(run)
{:error, %Ecto.Changeset{}}

get_dataclip(run)

@spec get_dataclip(run_or_uuid :: Lightning.Invocation.Run.t() | Ecto.UUID.t()) ::
  Lightning.Invocation.Dataclip.t() | nil

Gets a single dataclip given one of:

a Dataclip uuid

a Run model, that has an associated dataclip via it's event

Returns nil if the Dataclip does not exist.

Examples

iex> get_dataclip("27b73932-16c7-4a72-86a3-85d805ccff98")
%Dataclip{}

iex> get_dataclip("27b73932-16c7-4a72-86a3-85d805ccff98")
nil

iex> get_dataclip(%Run{id: "a uuid"})
%Dataclip{}

get_dataclip!(id)

@spec get_dataclip!(id :: Ecto.UUID.t()) :: Lightning.Invocation.Dataclip.t()

Gets a single dataclip.

Raises Ecto.NoResultsError if the Dataclip does not exist.

Examples

iex> get_dataclip!(123)
%Dataclip{}

iex> get_dataclip!(456)
** (Ecto.NoResultsError)

get_dataclip_query(run)

Query for retrieving the dataclip that a runs starting dataclip.

get_result_dataclip_query(run)

Query for retrieving the dataclip that was the result of a successful run.

get_run!(id)

@spec get_run!(Ecto.UUID.t() | Lightning.Invocation.Event.t()) ::
  Lightning.Invocation.Run.t()

Gets a single run.

Raises Ecto.NoResultsError if the Run does not exist.

Examples

iex> get_run!(123)
%Run{}

iex> get_run!(456)
** (Ecto.NoResultsError)

get_run_with_job!(id)

Fetches a run and preloads the job via the run's event.

list_dataclips()

@spec list_dataclips() :: [Lightning.Invocation.Dataclip.t()]

Returns the list of dataclips.

Examples

iex> list_dataclips()
[%Dataclip{}, ...]

list_dataclips(project)

@spec list_dataclips(project :: Lightning.Projects.Project.t()) :: [
  Lightning.Invocation.Dataclip.t()
]

list_dataclips_query(project)

@spec list_dataclips_query(project :: Lightning.Projects.Project.t()) ::
  Ecto.Queryable.t()

list_runs()

Returns the list of runs.

Examples

iex> list_runs()
[%Run{}, ...]

list_runs_for_project(project, params \\ %{})

update_dataclip(dataclip, attrs)

Updates a dataclip.

Examples

iex> update_dataclip(dataclip, %{field: new_value})
{:ok, %Dataclip{}}

iex> update_dataclip(dataclip, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_run(run, attrs)

Updates a run.

Examples

iex> update_run(run, %{field: new_value})
{:ok, %Run{}}

iex> update_run(run, %{field: bad_value})
{:error, %Ecto.Changeset{}}