Lightning.Invocation.Dataclip (Lightning v0.5.0)

Ecto model for Dataclips.

Dataclips represent some data that arrived in the system, and records both the data and the source of the data.

Types

:http_request The data arrived via a webhook.

:global Was created manually, and is intended to be used multiple times. When repetitive static data is needed to be maintained, instead of hard-coding into a Job - a more convenient solution is to create a :global Dataclip and access it inside the Job.

:run_result The final state of a successful run.

:saved_input An arbitrary input, created by a user. (Only configuration will be overwritten.)

Types

source_type()

@type source_type() :: :http_request | :global | :run_result | :saved_input

t()

@type t() :: %Lightning.Invocation.Dataclip{
  __meta__: Ecto.Schema.Metadata.t(),
  body: %{} | nil,
  id: Ecto.UUID.t() | nil,
  inserted_at: term(),
  project: term(),
  project_id: Ecto.UUID.t() | nil,
  source_run:
    Lightning.Invocation.Run.t() | Ecto.Association.NotLoaded.t() | nil,
  type: term(),
  updated_at: term()
}

Functions

get_types()

new(attrs \\ %{})

validate_by_type(changeset)

Append validations based on the type of the Dataclip.

:run_result must have an associated Run model.