Lightning.Jobs (Lightning v0.1.12)

The Jobs context.

Functions

change_job(job, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking job changes.

Examples

iex> change_job(job)
%Ecto.Changeset{data: %Job{}}

create_job(attrs \\ %{})

Creates a job.

Examples

iex> create_job(%{field: value})
{:ok, %Job{}}

iex> create_job(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_job(job)

Deletes a job.

Examples

iex> delete_job(job)
{:ok, %Job{}}

iex> delete_job(job)
{:error, %Ecto.Changeset{}}

get_downstream_jobs_for(job, trigger_type \\ nil)

@spec get_downstream_jobs_for(
  Lightning.Jobs.Job.t() | Ecto.UUID.t(),
  Lightning.Jobs.Trigger.trigger_type() | nil
) :: [Lightning.Jobs.Job.t()]

Returns the list of downstream jobs for a given job, optionally matching a specific trigger type.

get_job!(id)

Gets a single job.

Raises Ecto.NoResultsError if the Job does not exist.

Examples

iex> get_job!(123)
%Job{}

iex> get_job!(456)
** (Ecto.NoResultsError)

get_job(id)

get_job_by_webhook(path)

Gets a single job basic on it's webhook trigger.

get_jobs_for_cron_execution(timestamp)

@spec get_jobs_for_cron_execution(integer()) :: [Lightning.Jobs.Job.t()]

Returns a list of jobs to execute, given a current timestamp in Unix. This is used by the scheduler, which calls this function once every minute.

get_upstream_jobs_for(job)

@spec get_upstream_jobs_for(Lightning.Jobs.Job.t()) :: [Lightning.Jobs.Job.t()]

Returns the list of jobs excluding the one given.

get_workflow_for(job)

get_workflows_for(project)

Retrieves a list of Jobs with their initiating Job id in a tuple. This is useful when needing to group a projects jobs by their 'Workflow'

E.g. {<initial_job_uuid>, %Job{}}

jobs_for_project(project)

@spec jobs_for_project(Lightning.Projects.Project.t()) :: [Lightning.Jobs.Job.t()]

jobs_for_project_query(project)

@spec jobs_for_project_query(Lightning.Projects.Project.t()) :: Ecto.Queryable.t()

list_active_cron_jobs()

list_jobs()

Returns the list of jobs.

update_job(job, attrs)

Updates a job.

Examples

iex> update_job(job, %{field: new_value})
{:ok, %Job{}}

iex> update_job(job, %{field: bad_value})
{:error, %Ecto.Changeset{}}