Lightning.Jobs.Job (Lightning v0.2.0)

Ecto model for Jobs.

A Job contains the fields for defining a job.

body The expression/javascript code

name A plain text identifier

adaptor An NPM style string that contains both the module name and it's version. E.g. @openfn/language-http@v1.2.3 or @openfn/language-foo@latest. While the version suffix isn't enforced here as it's not strictly necessary in this context, the front end will ensure a version is stated (@latest being the default).

trigger Association to it's trigger, a job must have a trigger. See Lightning.Jobs.Trigger.

Types

t()

@type t() :: %Lightning.Jobs.Job{
  __meta__: Ecto.Schema.Metadata.t(),
  adaptor: String.t() | nil,
  body: String.t() | nil,
  credential:
    nil | Lightning.Credentials.Credential.t() | Ecto.Association.NotLoaded.t(),
  enabled: boolean(),
  events: term(),
  id: Ecto.UUID.t() | nil,
  inserted_at: term(),
  name: String.t() | nil,
  project:
    nil | Lightning.Projects.Project.t() | Ecto.Association.NotLoaded.t(),
  project_credential: term(),
  project_credential_id: term(),
  project_id: term(),
  trigger: nil | Lightning.Jobs.Trigger.t() | Ecto.Association.NotLoaded.t(),
  updated_at: term(),
  workflow:
    nil | Lightning.Workflows.Workflow.t() | Ecto.Association.NotLoaded.t(),
  workflow_id: term()
}

Functions

add_to_workflow(changeset)

@spec add_to_workflow(Ecto.Changeset.t()) :: Ecto.Changeset.t()

When a job is saved (created or updated) its workflow must be set. (All jobs belong to a workflow.)

If the job's trigger is of type :on_job_success or :on_job_failure, this function will set the workflow_id to that of its upstream job.

If the job's trigger is of type :cron or :webhook, it is the most upstream job in a workflow and this function will ensure that the job is in its own workflow—either creating a new one or leaving it unchanged.