Lightning.Jobs.Query (Lightning v0.4.9)

Query module for finding Jobs.

Functions

enabled_cron_jobs()

@spec enabled_cron_jobs() :: Ecto.Queryable.t()

Returns active jobs with their cron triggers for use in the cron scheduling service.

jobs_for(user)

@spec jobs_for(Lightning.Accounts.User.t()) :: Ecto.Queryable.t()

@spec jobs_for(Lightning.Projects.Project.t()) :: Ecto.Queryable.t()

Returns all jobs accessible to a user, via their projects or all jobs in a given project.