Lightning.KafkaTriggers (Lightning v2.7.14)

Contains the logic to manage kafka trigger and their associated pipelines.

Functions

build_topic_partition_offset(message)

Generate the key that is used to identify duplicate messages when used in association with the trigger id.

convert_rate_limit()

determine_offset_reset_policy(trigger)

Selects the appropriate offset reset policy for a given trigger based on the presence of partition-specific timestamps.

enable_disable_triggers(triggers)

find_enabled_triggers()

generate_pipeline_child_spec(trigger)

Generate the child spec needed to start a Pipeline child process.

get_kafka_triggers_being_updated(changeset)

start_triggers()

update_pipeline(supervisor, trigger_id)