Lightning.Pipeline.Runner (Lightning v0.1.13)

Job running entrypoint

Functions

create_dataclip_from_result(result, run)

@spec create_dataclip_from_result(
  result :: Engine.Result.t(),
  run :: Lightning.Invocation.Run.t()
) :: {:ok, Lightning.Invocation.Dataclip.t()} | {:error, any()}

Creates a dataclip linked to the run that just finished. If either the file doesn't exist or there is a JSON decoding error, it logs and returns an error tuple.

find_or_install_adaptor(adaptor)

@spec find_or_install_adaptor(adaptor :: String.t()) :: Engine.Adaptor.t()

Make sure an adaptor matching the name is available.

If it is available, return it's Engine.Adaptor struct - if not then install it.

start(run, opts \\ [])

@spec start(run :: Lightning.Invocation.Run.t(), opts :: []) :: Engine.Result.t()

Execute a Run.

Given a valid run:

Persist the Dataclip and the Job's body to disk

Create a blank output file on disk

Build up a %Engine.Runspec{} with the paths, and adaptor module name

And start it via Handler.start/2.

The callbacks implemented on Handler (c:Handler.on_start/1 and c:Handler.on_finish/2) update the run when a Run is started and when it's finished, attaching the exit_code and log when they are available.