Lightning.Pipeline.StateAssembler (Lightning v0.1.5)

Facilities for building the state for a Run

How state is assembled

For the most common jobs, where a Webhooks will store an :http_request type dataclip. The event that was created is associated with the dataclip.

At runtime, the initial state for a Run will in the shape of:

{ "data": <the dataclip>, "configuration": <the job's credential> }

Flow Jobs

When a Job is triggered by a previous Jobs success or failure these are the rules for constructing that Jobs state.

For jobs that trigger on it's upstream jobs failure, the event will have the source events original dataclip as its dataclip.

The state will also have the log of the source events run attached on the error key.

For Jobs triggered by a previous success, the event will have the source event runs resulting dataclip as its dataclip.

:run_result dataclips are expected to already have a data key, and are merged into the root.

Functions

assemble(run)

Assemble state for use in a Run.

context_query(map)