Lightning.Projects.File (Lightning v2.15.6)

The project files module

Types

t()

@type t() :: %Lightning.Projects.File{
  __meta__: Ecto.Schema.Metadata.t(),
  created_by: Lightning.Accounts.User.t(),
  created_by_id: term(),
  id: Ecto.UUID.t() | nil,
  inserted_at: term(),
  path: String.t(),
  project: Lightning.Projects.Project.t(),
  project_id: term(),
  size: term(),
  status: atom(),
  type: atom(),
  updated_at: term()
}

Functions

mark_failed(file)

@spec mark_failed(t()) :: Ecto.Changeset.t()

new(attrs)

@spec new(%{required(atom()) => any()}) :: Ecto.Changeset.t()

Creates a new file changeset.

Unlike regular changeset functions, this expects atom keys and the project and user to be passed in as models.