Lightning.Run (Lightning v2.15.8)

Ecto model for Runs.

Types

t()

@type t() :: %Lightning.Run{
  __meta__: Ecto.Schema.Metadata.t(),
  claimed_at: term(),
  created_by: term(),
  created_by_id: term(),
  dataclip: term(),
  dataclip_id: term(),
  error_type: term(),
  finished_at: term(),
  id: Ecto.UUID.t() | nil,
  inserted_at: term(),
  log_lines: term(),
  options: term(),
  priority: term(),
  run_steps: term(),
  snapshot: term(),
  snapshot_id: term(),
  started_at: term(),
  starting_job: term(),
  starting_job_id: term(),
  starting_trigger: term(),
  starting_trigger_id: term(),
  state: term(),
  steps: term(),
  updated_at: term(),
  work_order: Lightning.WorkOrder.t() | Ecto.Association.NotLoaded.t(),
  work_order_id: term(),
  worker_name: term(),
  workflow: term()
}

Functions

complete(run, params)

@spec complete(
  %Lightning.Run{
    __meta__: term(),
    claimed_at: term(),
    created_by: term(),
    created_by_id: term(),
    dataclip: term(),
    dataclip_id: term(),
    error_type: term(),
    finished_at: term(),
    id: term(),
    inserted_at: term(),
    log_lines: term(),
    options: term(),
    priority: term(),
    run_steps: term(),
    snapshot: term(),
    snapshot_id: term(),
    started_at: term(),
    starting_job: term(),
    starting_job_id: term(),
    starting_trigger: term(),
    starting_trigger_id: term(),
    state: term(),
    steps: term(),
    updated_at: term(),
    work_order: term(),
    work_order_id: term(),
    worker_name: term(),
    workflow: term()
  },
  %{optional(any()) => any()}
) :: Ecto.Changeset.t()

failure_states()

Returns the list of failure states for a run.

These are all final states except :success.

final_states()

Returns the list of final states for a run.

for(trigger, attrs)

new(attrs \\ %{})

start(run, params)