Lightning.Runs.PromExPlugin (Lightning v2.7.10)

Metrics callbacks implementation for the PromEx plugin.

Event metrics are used to publish the dispatched events in the system. Polling metrics are used to publish metrics that are calculated by polling data periodically.

Functions

calculate_average_claim_duration(reference_time, run_age_seconds)

count_finalised_runs(threshold_time)

run_queue_metrics(run_age_seconds, consideration_window_seconds)

stalled_run_count(threshold_seconds)