Lightning.Runtime.RuntimeManager (Lightning v2.0.0-rc10)

Locates and runs the Runtime server. Added in order to ease development and default installations of Lightning

https://registry.npmjs.org/lightning-runtime/latest

Runtime configuration

Sample:

config :lightining, Elixir.Lightning.Runtime.RuntimeManager,

version: "0.1.0",
start: true,
args: ~w(js/app.js --bundle --target=es2016 --outdir=../priv/static/assets),
cd: Path.expand("../assets", __DIR__),
env: %{}

Options:

:version - the expected runtime version

:start - flag to start the runtime manager. If false the GenServer won't be started

:path - the path to find the runtime executable at. By default, it is automatically downloaded and placed inside the _build directory of your current app

Overriding the :path is not recommended, as we will automatically download and manage the runtime for you.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

start_link(args)