Lightning.Scrubber (Lightning v0.6.0)

Process used to scrub strings of sensitive information.

Can be started via start_link/1.

{:ok, scrubber} =
  Lightning.Scrubber.start_link(
    samples:
      Lightning.Credentials.sensitive_values_for(credential)
  )

Takes an optional :name key, in case you need to name the process.

Functions

child_spec(arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

encode_samples(samples)

@spec encode_samples(samples :: [String.t()]) :: [String.t()]

Prepare a list of sensitive samples (strings) into a potentially bigger list composed of variations a sample may appear.

samples(agent)

scrub(agent, lines)

start_link(opts)

@spec start_link(opts :: [samples: [String.t()], name: nil | GenServer.name()]) ::
  Agent.on_start()