Lightning.WebhookAuthMethods (Lightning v0.10.0-pre4)

The Lightning.WebhookAuthMethods context provides a suite of functionalities to handle and interact with Webhook Authentication Methods within the application.

Overview

This context is designed to encapsulate all the logic related to Webhook Authentication Methods, including their creation, updates, retrievals, and deletions. It serves as an API for other parts of the system to interact with Webhook Authentication Methods, ensuring that the internal implementation details are hidden, and only the necessary functionalities are exposed.

Functionalities

create_auth_method/1: Allows for the creation of a new Webhook Authentication Method with given attributes.

update_auth_method/2: Facilitates the updating of an existing Webhook Authentication Method.

delete_auth_method/1: Allows for the deletion of a given Webhook Auth Method.

list_for_project/1: Provides a list of all Webhook Authentication Methods for a given project.

list_for_trigger/1: Provides a list of all Webhook Authentication Methods for a given trigger.

find_by_api_key/2: Retrieves a Webhook Authentication Method by API key and project_id.

find_by_username_and_password/3: Obtains a Webhook Authentication Method by username, password, and project_id if the password is valid.

find_by_id!/2: Fetches a Webhook Authentication Method by id and project_id, raises if it does not exist.

Examples

The examples in each function provide detailed use cases and expected outcomes, aiding developers in understanding how to utilize the functionalities provided by this context.

Usage

This module is intended to be used by other modules, contexts, or controllers that need to perform operations related to Webhook Authentication Methods, serving as a clear and consistent interface for these operations.

Functions

create_auth_method(attrs, list)

Creates a WebhookAuthMethod.

Examples

iex> create_auth_method(%{valid attributes})
{:ok, %WebhookAuthMethod{}}

iex> create_auth_method(%{invalid attributes})
{:error, %Ecto.Changeset{}}

create_auth_method(trigger, params, list)

@spec create_auth_method(Lightning.Workflows.Trigger.t(), map(), [
  {:actor, Lightning.Accounts.User.t()}
]) ::
  {:ok, Lightning.Workflows.WebhookAuthMethod.t()}
  | {:error, Ecto.Changeset.t()}

create_changeset(webhook_auth_method, params)

delete_auth_method(auth_method)

Deletes a Webhook Auth Method.

Examples

iex> delete_auth_method(%WebhookAuthMethod{id: "some_id"})
{:ok, %WebhookAuthMethod{}}

iex> delete_auth_method(%WebhookAuthMethod{id: "non_existing_id"})
{:error, :conflict}

find_by_api_key(api_key, project)

Gets a WebhookAuthMethod by api_key and project_id.

Examples

iex> find_by_api_key("existing_api_key", %Project{id: "existing_project_id"})
%WebhookAuthMethod{}

iex> find_by_id_by_username("non_existing_api_key", %Project{id: "existing_project_id"})
nil

find_by_id!(id)

Gets a WebhookAuthMethod by id and project_id, raises if it does not exist.

Examples

iex> find_by_id!("existing_id", %Project{id: "existing_project_id"})
%WebhookAuthMethod{}

iex> find_by_id!("non_existing_id", %Project{id: "existing_project_id"})
** (Ecto.NoResultsError)

find_by_username_and_password(username, password, project)

Gets a WebhookAuthMethod by username, password, and project_id if the password is valid.

Examples

iex> find_by_username_and_password("existing_username", "valid_password", %Project{id: "existing_project_id"})
%WebhookAuthMethod{}

iex> find_by_username_and_password("existing_username", "invalid_password", %Project{id: "existing_project_id"})
nil

list_for_project(project)

Lists all WebhookAuthMethods for a given project.

Examples

iex> list_for_project(%Project{id: "existing_project_id"})
[%WebhookAuthMethod{}, ...]

iex> list_for_project(%Project{id: "non_existing_project_id"})
[]

list_for_trigger(trigger)

Retrieves a list of webhook_auth_methods associated with the given trigger.

Parameters

trigger: The Trigger struct for which associated webhook_auth_methods are to be retrieved. If nil, the function will return nil.

Returns

Returns a list of %WebhookAuthMethod{} associated with the given trigger.

Examples

iex> Lightning.Workflows.list_for_trigger(trigger)
[%WebhookAuthMethod{}, ...]

perform(job)

When called with %{"type" => "purge_deleted"}, finds webhook_auth_methods that are ready for permanent deletion, disassociates them from the triggers, and then deletes the webhook_auth_methods.

schedule_for_deletion(webhook_auth_method)

Schedules a given webhook auth method for deletion.

The deletion date is determined based on the :purge_deleted_after_days configuration in the application environment. If this configuration is absent, the webhook auth method is scheduled for immediate deletion.

The function will also perform necessary side effects such as:

Removing associations of the webhook auth method.

Notifying the owner of the webhook auth method about the scheduled deletion.

Parameters

webhook_auth_method: A WebhookAuthMethod struct that is to be scheduled for deletion.

Returns

{:ok, webhook_auth_method}: Returns an :ok tuple with the updated webhook auth method struct if the update was successful.

{:error, changeset}: Returns an :error tuple with the changeset if the update failed.

Examples

iex> schedule_for_deletion(%WebhookAuthMethod{id: some_id})
{:ok, %WebhookAuthMethod{}}

iex> schedule_for_deletion(%WebhookAuthMethod{})
{:error, %Ecto.Changeset{}}

update_auth_method(webhook_auth_method, attrs, list)

Updates a WebhookAuthMethod.

Examples

iex> update_auth_method(webhook_auth_method, %{field: new_value})
{:ok, %WebhookAuthMethod{}}

iex> update_auth_method(webhook_auth_method, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_trigger_auth_methods(trigger, auth_methods, list)

@spec update_trigger_auth_methods(
  Lightning.Workflows.Trigger.t(),
  [Lightning.Workflows.WebhookAuthMethod.t(), ...] | [],
  [{:actor, Lightning.Accounts.User.t()}]
) :: {:ok, Lightning.Workflows.Trigger.t()} | {:error, Ecto.Changeset.t()}