Lightning.WorkOrders (Lightning v0.10.2)

Context for creating WorkOrders.

Workorders

Workorders represent the entrypoint for a unit of work in Lightning. They allow you to track the status of a webhook or cron trigger.

For example if a user makes a request to a webhook endpoint, a Workorder is created with it's associated Workflow and Dataclip.

Every Workorder has at least one Attempt, which represents a single invocation of the Workflow. If the workflow fails, and the attempt is retried, a new Attempt is created on the Workorder.

This allows you group all the attempts for a single webhook, and track the success or failure of a given dataclip.

Creating Workorders

Workorders can be created in three ways:

Via a webhook trigger

Via a cron trigger

Manually by a user (via the UI or API)

Retries do not create new Workorders, but rather new Attempts on the existing Workorder.

Types

work_order_option()

@type work_order_option() ::
  {:workflow, Lightning.Workflows.Workflow.t()}
  | {:dataclip, Lightning.Invocation.Dataclip.t()}
  | {:created_by, Lightning.Accounts.User.t()}

Functions

build_for(trigger, attrs)

@spec build_for(Lightning.Workflows.Trigger.t() | Lightning.Workflows.Job.t(), map()) ::
  Ecto.Changeset.t(Lightning.WorkOrder.t())

create_for(manual)

create_for(trigger, opts)

@spec create_for(Lightning.Workflows.Trigger.t() | Lightning.Workflows.Job.t(), [
  work_order_option()
]) ::
  {:ok, Lightning.WorkOrder.t()}
  | {:error, Ecto.Changeset.t(Lightning.WorkOrder.t())}

Create a new Workorder.

For a webhook

create_for(trigger, workflow: workflow, dataclip: dataclip)

For a user

create_for(job, workflow: workflow, dataclip: dataclip, user: user)

get(id, opts \\ [])

@spec get(Ecto.UUID.t(), [{:include, [atom()]}]) :: Lightning.WorkOrder.t() | nil

Get a Workorder by id.

Optionally preload associations by passing a list of atoms to :include.

Lightning.WorkOrders.get(id, include: [:attempts])

retry(attempt, run, opts \\ [])

@spec retry(
  Lightning.Attempt.t() | Ecto.UUID.t(),
  Lightning.Invocation.Run.t() | Ecto.UUID.t(),
  [work_order_option(), ...] | []
) ::
  {:ok, Lightning.Attempt.t()}
  | {:error, Ecto.Changeset.t(Lightning.Attempt.t())}

Retry an Attempt from a given run.

This will create a new Attempt on the Workorder, and enqueue it for processing.

When creating a new Attempt, a graph of the workflow is created, and using that graph the runs would not be replaced with the new runs are linked.

For example, by retrying a run from the middle of the workflow, the new attempt will only contain the runs that are upstream of the run being retried.

retry_many(workorders, opts)

@spec retry_many(
  [Lightning.WorkOrder.t(), ...] | [Lightning.AttemptRun.t(), ...],
  [work_order_option(), ...] | []
) :: {:ok, count :: integer()}

retry_many(workorders, job_id, opts)

@spec retry_many(
  [Lightning.WorkOrder.t(), ...],
  job_id :: Ecto.UUID.t(),
  [work_order_option(), ...] | []
) :: {:ok, count :: integer()}

subscribe(project_id)

See Lightning.WorkOrders.Events.subscribe/1.

update_state(attempt)

@spec update_state(Lightning.Attempt.t()) :: {:ok, Lightning.WorkOrder.t()}

Updates the state of a WorkOrder based on the state of an attempt.

This considers the state of all attempts on the WorkOrder, with the Attempt passed in as the latest attempt.

See Lightning.WorkOrders.Query.state_for/1 for more details.