Lightning.WorkOrders.SearchParams (Lightning v2.6.0)

This module is used to parse search parameters for work orders and provide a query to the database.

Types

t()

@type t() :: %Lightning.WorkOrders.SearchParams{
  date_after: DateTime.t(),
  date_before: DateTime.t(),
  search_fields: [String.t()],
  search_term: String.t(),
  status: [String.t()],
  wo_date_after: DateTime.t(),
  wo_date_before: DateTime.t(),
  workflow_id: Ecto.UUID.t(),
  workorder_id: term()
}

Functions

all_statuses_set?(map)

new(params)

status_list()
(macro)

to_uri_params(search_params)