Lightning.Workflows.Edge (Lightning v2.9.2)

Ecto model for Workflow Edges.

A Workflow Edge represents a connection between two jobs (or a trigger and a job) in a workflow.

The source of the edge is either a job or a trigger. The target of the edge is always a job.

Types

edge_condition()

@type edge_condition() :: :always | :on_job_success | :on_job_failure | :js_expression

t()

@type t() :: %Lightning.Workflows.Edge{
  __meta__: Ecto.Schema.Metadata.t(),
  condition_expression: term(),
  condition_label: term(),
  condition_type: edge_condition(),
  delete: boolean(),
  enabled: boolean(),
  id: Ecto.UUID.t() | nil,
  inserted_at: term(),
  source_job:
    nil | Lightning.Workflows.Job.t() | Ecto.Association.NotLoaded.t(),
  source_job_id: term(),
  source_trigger:
    nil | Lightning.Workflows.Trigger.t() | Ecto.Association.NotLoaded.t(),
  source_trigger_id: term(),
  target_job:
    nil | Lightning.Workflows.Job.t() | Ecto.Association.NotLoaded.t(),
  target_job_id: term(),
  updated_at: term(),
  workflow:
    nil | Lightning.Workflows.Workflow.t() | Ecto.Association.NotLoaded.t(),
  workflow_id: term()
}

Functions

changeset(edge, attrs)

new(attrs)

validate(changeset)