Lightning.Workflows.Snapshot (Lightning v2.15.0-pre4)

Ecto model for Workflow Snapshots.

Snapshots are a way to store the state of a workflow at a given point in time.

Types

t()

@type t() :: %Lightning.Workflows.Snapshot{
  __meta__: Ecto.Schema.Metadata.t(),
  edges: [
    %Lightning.Workflows.Snapshot.Edge{
      condition_expression: term(),
      condition_label: term(),
      condition_type: term(),
      enabled: term(),
      id: term(),
      inserted_at: term(),
      source_job_id: term(),
      source_trigger_id: term(),
      target_job_id: term(),
      updated_at: term()
    }
  ],
  id: Ecto.UUID.t() | nil,
  inserted_at: term(),
  jobs: [
    %Lightning.Workflows.Snapshot.Job{
      adaptor: term(),
      body: term(),
      credential: term(),
      id: term(),
      inserted_at: term(),
      keychain_credential: term(),
      keychain_credential_id: term(),
      name: term(),
      project_credential: term(),
      project_credential_id: term(),
      updated_at: term()
    }
  ],
  lock_version: term(),
  name: String.t() | nil,
  positions: map() | nil,
  triggers: [
    %Lightning.Workflows.Snapshot.Trigger{
      comment: term(),
      cron_expression: term(),
      custom_path: term(),
      enabled: term(),
      has_auth_method: term(),
      id: term(),
      inserted_at: term(),
      kafka_configuration: term(),
      type: term(),
      updated_at: term(),
      webhook_auth_methods: term(),
      webhook_reply: term()
    }
  ],
  workflow:
    nil | Lightning.Workflows.Workflow.t() | Ecto.Association.NotLoaded.t(),
  workflow_id: term()
}

Functions

build(workflow)

@spec build(Lightning.Workflows.Workflow.t()) :: Ecto.Changeset.t()

create(workflow)

@spec create(Lightning.Workflows.Workflow.t()) ::
  {:ok, t()} | {:error, Ecto.Changeset.t()}

get_all_by_ids(ids)

get_all_for(workflow)

@spec get_all_for(Lightning.Workflows.Workflow.t()) :: [t()]

get_by_version(workflow_id, version)

get_current_for(workflow)

@spec get_current_for(Lightning.Workflows.Workflow.t()) :: t() | nil

Get the latest snapshot for a workflow, based on the lock_version.

It returns the latest snapshot regardless of the lock_version of the workflow passed in.

include_latest_snapshot(multi, name \\ :snapshot, workflow)

@spec include_latest_snapshot(
  Ecto.Multi.t(),
  binary() | :snapshot,
  Lightning.Workflows.Workflow.t()
) ::
  Ecto.Multi.t()

new(workflow)