Lightning.Workflows.Triggers.KafkaConfiguration (Lightning v2.7.11)

Configuration of Kafka Triggers.

Functions

apply_hosts_string(changeset)

apply_topics_string(changeset)

changeset(kafka_configuration, attrs)

generate_hosts_string(changeset)

generate_topics_string(changeset)

partitions_changeset(configuration, partition, timestamp)

Returns a changeset to maintain persisted partition timestamps. These timestamps are used to provide an updated offset reset policy should the associated consumer group have been used previously but has not connected to the cluster for a long enough time that the cluster no longer has a committed offset.

sasl_types()

set_group_id_if_required(changeset)

validate_initial_offset_reset_policy(changeset)

validate_sasl_credentials(changeset)