LightningWeb.Components.Icon (Lightning v0.1.7)

For Lightning-specific concepts, we define icons here for ease of management and reuse.

Note: we're in the process of migrating away from defining SVGs here and instead using Heroicons from Petal.

Functions

chevron_left(assigns)

chevron_right(assigns)

dataclips(assigns)

eye(assigns)

left(assigns)

plus(assigns)

right(assigns)

runs(assigns)

trash(assigns)

workflows(assigns)