LightningWeb.Components.Modal (Lightning v2.15.1)

A modal component that can be used to display a modal on the page.

This currently isn't used anywhere but should be used in the future to replace the existing modal implementations.

Functions

hide_modal(js \\ %JS{}, id)

modal(assigns)

Attributes

id (:string) (required)

show (:boolean) - Defaults to false.

position (:string) - Defaults to "fixed inset-0".

width (:string) - Defaults to "max-w-3xl".

close_on_click_away (:boolean) - Defaults to true.

close_on_keydown (:boolean) - Defaults to true.

on_close (Phoenix.LiveView.JS) - Defaults to %Phoenix.LiveView.JS{ops: []}.

on_open (Phoenix.LiveView.JS) - Defaults to %Phoenix.LiveView.JS{ops: []}.

Global attributes are accepted.

Slots

inner_block (required)

title

subtitle

footer - Accepts attributes:

class (:string)

modal_footer(assigns)

Attributes

class (:any) - Defaults to "sm:flex sm:flex-row-reverse gap-3".

Slots

inner_block (required)

modal_title(assigns)

Attributes

Global attributes are accepted.

Slots

subtitle

inner_block (required)

show_modal(js \\ %JS{}, id)