LightningWeb.Components.Viewers (Lightning v0.12.1)

Components for rendering Logs and Dataclips.

Scrolling can be tricky

We seldom know how long a log or a dataclip will be, and we want to be able to contain the element in a fixed height container. In some situations wrapping the component in a div with inline-flex class will help with scrolling.

Functions

dataclip_viewer(assigns)

Attributes

id (:string) (required)

stream (:list) (required) - A stream of lines to render. In the shape of %{id: String.t(), line: String.t(), index: integer()}.

class (:string) - Additional classes to add to the log viewer container. Defaults to nil.

type (:atom) - Defaults to nil.

log_viewer(assigns)

Renders out a log line stream

Internally it uses the LogLineHighlight hook to highlight the log line with the highlight_id attribute.

Example

<Viewers.log_viewer
  id="log-viewer-data"
  stream={@log_lines}
  highlight_id={@selected_run_id}
/>

Attributes

id (:string) (required)

stream (:list) (required) - A stream of Lightning.Invocation.LogLine structs.

highlight_id (:string) - The id of the log line to highlight, matching the run_id field. Defaults to nil.

class (:string) - Additional classes to add to the log viewer container. Defaults to nil.