LightningWeb.CredentialLive.GoogleSheetsComponent (Lightning v0.11.0)

Form component to setup a Google Sheets component.

This component has several moving parts:

Subscribes to a PubSub topic specially link to the component id See: LightningWeb.OauthCredentialHelper.

Uses the Lightning.Google module to set up an OAuth client for generating urls, exchanging the code and requesting a new access_token.

The flow for creating a new token is:

Generate an authorization link which contains:

The authorization url from the Google client with the applications callback_url

A state string that is an encrypted set of data with the components module and id in it

Once the user authorizes the client the callback is requested with a code

The LightningWeb.OidcController decodes the state returned to it and does a 'broadcast_forward' which is simply a message expected to be received by a LiveView and applied to Phoenix.LiveView.send_update/3.

The component receives the code and requests a token.

Any changes to the token (Credential body) are still handled by the parent component and so a update_body function is passed in to send params changes back up to update the form.

Functions

authorize_button(assigns)

disabled_authorize_button(assigns)

error_block(assigns)

fieldset(assigns)

Attributes

form (:map) (required)

update_body (:any) (required)

Slots

inner_block

userinfo(assigns)