LightningWeb.OauthCredentialHelper (Lightning v2.4.14)

A set of helper functions to encodes state and coordinate OAuth callbacks back to a LiveView component.

Functions

broadcast(subscription_id, msg)

broadcast_forward(subscription_id, mod, opts)

Broadcast a message specifically for forwarding a message to a component. It expects a subscription_id, the module of the component and opts being a keyword list containing an :id key of the specific component.

See: Phoenix.LiveView.send_update/3 for more info.

A corresponding LiveView (that is subscribed) is expected to have a matching handle_info/2 that looks like this:

def handle_info({:forward, mod, opts}, socket) do
  send_update(mod, opts)
  {:noreply, socket}
end

build_state(subscription_id, mod, component_id)

Encode and encrypt the callback data which will be sent so a provider as the state key in the request.

The values are:

subscription_id The same ID used to subscribe.

The component module The LiveView component that is going to receive update

The component id The ID of the component

decode_state(state)

subscribe(subscription_id)

Subscribe to the oauth_credential topic. It expects the a unique ID for the topic, usually the LiveView's socket.id.