LightningWeb.PageView (Lightning v2.15.2)

Functions

__resource__()

The resource name, as an atom, for this view

render(template, assigns \\ %{})

Renders the given template locally.

template_not_found(template, assigns)

@spec template_not_found(binary(), map()) :: no_return()

Callback invoked when no template is found. By default it raises but can be customized to render a particular template.