LightningWeb.Router.Helpers (Lightning v0.1.13)

Module with named helpers generated from LightningWeb.Router.

Functions

api_job_path(conn_or_endpoint, action)

api_job_path(conn_or_endpoint, action, params)

api_job_path(conn_or_endpoint, action, id, params)

api_job_url(conn_or_endpoint, action)

api_job_url(conn_or_endpoint, action, params)

api_job_url(conn_or_endpoint, action, id, params)

api_project_job_path(conn_or_endpoint, action, project_id)

api_project_job_path(conn_or_endpoint, action, project_id, params)

api_project_job_path(conn_or_endpoint, action, project_id, id, params)

api_project_job_url(conn_or_endpoint, action, project_id)

api_project_job_url(conn_or_endpoint, action, project_id, params)

api_project_job_url(conn_or_endpoint, action, project_id, id, params)

api_project_path(conn_or_endpoint, action)

api_project_path(conn_or_endpoint, action, params)

api_project_path(conn_or_endpoint, action, id, params)

api_project_run_path(conn_or_endpoint, action, project_id)

api_project_run_path(conn_or_endpoint, action, project_id, params)

api_project_run_path(conn_or_endpoint, action, project_id, id, params)

api_project_run_url(conn_or_endpoint, action, project_id)

api_project_run_url(conn_or_endpoint, action, project_id, params)

api_project_run_url(conn_or_endpoint, action, project_id, id, params)

api_project_url(conn_or_endpoint, action)

api_project_url(conn_or_endpoint, action, params)

api_project_url(conn_or_endpoint, action, id, params)

api_run_path(conn_or_endpoint, action)

api_run_path(conn_or_endpoint, action, params)

api_run_path(conn_or_endpoint, action, id, params)

api_run_url(conn_or_endpoint, action)

api_run_url(conn_or_endpoint, action, params)

api_run_url(conn_or_endpoint, action, id, params)

audit_index_path(conn_or_endpoint, action)

audit_index_path(conn_or_endpoint, action, params)

audit_index_url(conn_or_endpoint, action)

audit_index_url(conn_or_endpoint, action, params)

auth_providers_index_path(conn_or_endpoint, action)

auth_providers_index_path(conn_or_endpoint, action, params)

auth_providers_index_url(conn_or_endpoint, action)

auth_providers_index_url(conn_or_endpoint, action, params)

credential_edit_path(conn_or_endpoint, action)

credential_edit_path(conn_or_endpoint, action, params)

credential_edit_path(conn_or_endpoint, action, id, params)

credential_edit_url(conn_or_endpoint, action)

credential_edit_url(conn_or_endpoint, action, params)

credential_edit_url(conn_or_endpoint, action, id, params)

credential_index_path(conn_or_endpoint, action)

credential_index_path(conn_or_endpoint, action, params)

credential_index_url(conn_or_endpoint, action)

credential_index_url(conn_or_endpoint, action, params)

dashboard_index_path(conn_or_endpoint, action)

dashboard_index_path(conn_or_endpoint, action, params)

dashboard_index_url(conn_or_endpoint, action)

dashboard_index_url(conn_or_endpoint, action, params)

first_setup_superuser_path(conn_or_endpoint, action)

first_setup_superuser_path(conn_or_endpoint, action, params)

first_setup_superuser_url(conn_or_endpoint, action)

first_setup_superuser_url(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action)

live_dashboard_path(conn_or_endpoint, action, params)

live_dashboard_path(conn_or_endpoint, action, page, params)

live_dashboard_path(conn_or_endpoint, action, node, page, params)

live_dashboard_url(conn_or_endpoint, action)

live_dashboard_url(conn_or_endpoint, action, params)

live_dashboard_url(conn_or_endpoint, action, page, params)

live_dashboard_url(conn_or_endpoint, action, node, page, params)

oidc_path(conn_or_endpoint, action, provider)

oidc_path(conn_or_endpoint, action, provider, params)

oidc_url(conn_or_endpoint, action, provider)

oidc_url(conn_or_endpoint, action, provider, params)

path(data, path)

Generates the path information including any necessary prefix.

profile_edit_path(conn_or_endpoint, action)

profile_edit_path(conn_or_endpoint, action, params)

profile_edit_path(conn_or_endpoint, action, id, params)

profile_edit_url(conn_or_endpoint, action)

profile_edit_url(conn_or_endpoint, action, params)

profile_edit_url(conn_or_endpoint, action, id, params)

project_dataclip_edit_path(conn_or_endpoint, action, project_id)

project_dataclip_edit_path(conn_or_endpoint, action, project_id, params)

project_dataclip_edit_path(conn_or_endpoint, action, project_id, id, params)

project_dataclip_edit_url(conn_or_endpoint, action, project_id)

project_dataclip_edit_url(conn_or_endpoint, action, project_id, params)

project_dataclip_edit_url(conn_or_endpoint, action, project_id, id, params)

project_dataclip_index_path(conn_or_endpoint, action, project_id)

project_dataclip_index_path(conn_or_endpoint, action, project_id, params)

project_dataclip_index_url(conn_or_endpoint, action, project_id)

project_dataclip_index_url(conn_or_endpoint, action, project_id, params)

project_index_path(conn_or_endpoint, action)

project_index_path(conn_or_endpoint, action, params)

project_index_path(conn_or_endpoint, action, id, params)

project_index_url(conn_or_endpoint, action)

project_index_url(conn_or_endpoint, action, params)

project_index_url(conn_or_endpoint, action, id, params)

project_job_edit_path(conn_or_endpoint, action, project_id)

project_job_edit_path(conn_or_endpoint, action, project_id, params)

project_job_edit_path(conn_or_endpoint, action, project_id, id, params)

project_job_edit_url(conn_or_endpoint, action, project_id)

project_job_edit_url(conn_or_endpoint, action, project_id, params)

project_job_edit_url(conn_or_endpoint, action, project_id, id, params)

project_job_index_path(conn_or_endpoint, action, project_id)

project_job_index_path(conn_or_endpoint, action, project_id, params)

project_job_index_url(conn_or_endpoint, action, project_id)

project_job_index_url(conn_or_endpoint, action, project_id, params)

project_run_index_path(conn_or_endpoint, action, project_id)

project_run_index_path(conn_or_endpoint, action, project_id, params)

project_run_index_path(conn_or_endpoint, action, project_id, id, params)

project_run_index_url(conn_or_endpoint, action, project_id)

project_run_index_url(conn_or_endpoint, action, project_id, params)

project_run_index_url(conn_or_endpoint, action, project_id, id, params)

project_workflow_path(conn_or_endpoint, action, project_id)

project_workflow_path(conn_or_endpoint, action, project_id, params)

project_workflow_path(conn_or_endpoint, action, project_id, job_id, params)

project_workflow_url(conn_or_endpoint, action, project_id)

project_workflow_url(conn_or_endpoint, action, project_id, params)

project_workflow_url(conn_or_endpoint, action, project_id, job_id, params)

settings_index_path(conn_or_endpoint, action)

settings_index_path(conn_or_endpoint, action, params)

settings_index_url(conn_or_endpoint, action)

settings_index_url(conn_or_endpoint, action, params)

static_integrity(endpoint, path)

Generates an integrity hash to a static asset given its file path.

static_path(conn, path)

Generates path to a static asset given its file path.

static_url(conn, path)

Generates url to a static asset given its file path.

url(data)

Generates the connection/endpoint base URL without any path information.

user_confirmation_path(conn_or_endpoint, action)

user_confirmation_path(conn_or_endpoint, action, params)

user_confirmation_path(conn_or_endpoint, action, token, params)

user_confirmation_url(conn_or_endpoint, action)

user_confirmation_url(conn_or_endpoint, action, params)

user_confirmation_url(conn_or_endpoint, action, token, params)

user_edit_path(conn_or_endpoint, action)

user_edit_path(conn_or_endpoint, action, params)

user_edit_path(conn_or_endpoint, action, id, params)

user_edit_url(conn_or_endpoint, action)

user_edit_url(conn_or_endpoint, action, params)

user_edit_url(conn_or_endpoint, action, id, params)

user_index_path(conn_or_endpoint, action)

user_index_path(conn_or_endpoint, action, params)

user_index_path(conn_or_endpoint, action, id, params)

user_index_url(conn_or_endpoint, action)

user_index_url(conn_or_endpoint, action, params)

user_index_url(conn_or_endpoint, action, id, params)

user_registration_path(conn_or_endpoint, action)

user_registration_path(conn_or_endpoint, action, params)

user_registration_url(conn_or_endpoint, action)

user_registration_url(conn_or_endpoint, action, params)

user_reset_password_path(conn_or_endpoint, action)

user_reset_password_path(conn_or_endpoint, action, params)

user_reset_password_path(conn_or_endpoint, action, token, params)

user_reset_password_url(conn_or_endpoint, action)

user_reset_password_url(conn_or_endpoint, action, params)

user_reset_password_url(conn_or_endpoint, action, token, params)

user_session_path(conn_or_endpoint, action)

user_session_path(conn_or_endpoint, action, params)

user_session_path(conn_or_endpoint, action, token, params)

user_session_url(conn_or_endpoint, action)

user_session_url(conn_or_endpoint, action, params)

user_session_url(conn_or_endpoint, action, token, params)

webhooks_path(conn_or_endpoint, action, path)

webhooks_path(conn_or_endpoint, action, path, params)

webhooks_url(conn_or_endpoint, action, path)

webhooks_url(conn_or_endpoint, action, path, params)