LightningWeb.RunChannel (Lightning v2.15.7)

Phoenix channel to interact with Runs.

Functions

child_spec(init_arg)

handle_in(binary, payload, socket)

For the time being, calls to fetch:dataclip will return dataclips that are preformatted for use as "initial state" in a run.

This means that the body of http requests will be nested inside a "data" key.

There is an open discussion on the community that may impact how we store HTTP requests in the database as dataclips and how we send the body of those HTTP requests to the worker to use as initial state.

start_link(triplet)