LightningWeb.RunLive.Index (Lightning v2.9.10)

Index Liveview for Runs

Functions

checked?(changeset, id)

Takes a changeset used for querying workorders and checks to see if the given filter is present in that changeset. Returns true or false.

filters_changeset(params)

Creates a changeset based on given parameters and the fixed workorder filter types.

handle_info(arg1, socket)

When a WorkOrderCreated event is detected, we first check to see if the new work order is admissible on the page, given the current filters. If it is, we add it to the top of the page. If not, nothing happens.

render(assigns)

Callback implementation for Phoenix.LiveView.render/1.

validate_bulk_rerun(selected_work_orders, map)