mix lightning.install_schemas (Lightning v2.15.0-pre5)

Install the credential json schemas Use --exclude language-package1, language-package2 to exclude specific packages

Types

body()

@type body() :: HTTPoison.Base.body()

headers()

@type headers() :: HTTPoison.Base.headers()

method()

@type method() :: HTTPoison.Base.method()

options()

@type options() :: HTTPoison.Base.options()

params()

@type params() :: HTTPoison.Base.params()

request()

@type request() :: HTTPoison.Base.request()

url()

@type url() :: HTTPoison.Base.url()

Functions

delete(url, headers \\ [], options \\ [])

@spec delete(binary(), headers(), Keyword.t()) ::
  {:ok,
   HTTPoison.Response.t()
   | HTTPoison.AsyncResponse.t()
   | HTTPoison.MaybeRedirect.t()}
  | {:error, HTTPoison.Error.t()}

Issues a DELETE request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

delete!(url, headers \\ [], options \\ [])

@spec delete!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t()
  | HTTPoison.AsyncResponse.t()
  | HTTPoison.MaybeRedirect.t()

Issues a DELETE request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

fetch_schemas(excluded \\ [], fun)

get(url, headers \\ [], options \\ [])

@spec get(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues a GET request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

get!(url, headers \\ [], options \\ [])

@spec get!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues a GET request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

head(url, headers \\ [], options \\ [])

@spec head(binary(), headers(), Keyword.t()) ::
  {:ok, HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()}
  | {:error, HTTPoison.Error.t()}

Issues a HEAD request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

head!(url, headers \\ [], options \\ [])

@spec head!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t() | HTTPoison.AsyncResponse.t()

Issues a HEAD request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

options(url, headers \\ [], options \\ [])

@spec options(binary(), headers(), Keyword.t()) ::
  {:ok,
   HTTPoison.Response.t()
   | HTTPoison.AsyncResponse.t()
   | HTTPoison.MaybeRedirect.t()}
  | {:error, HTTPoison.Error.t()}

Issues an OPTIONS request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

options!(url, headers \\ [], options \\ [])

@spec options!(binary(), headers(), Keyword.t()) ::
  HTTPoison.Response.t()
  | HTTPoison.AsyncResponse.t()
  | HTTPoison.MaybeRedirect.t()

Issues a OPTIONS request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

parse_excluded(args)

patch(url, body, headers \\ [], options \\ [])

@spec patch(binary(), any(), headers(), Keyword.t()) ::
  {:ok,
   HTTPoison.Response.t()
   | HTTPoison.AsyncResponse.t()
   | HTTPoison.MaybeRedirect.t()}
  | {:error, HTTPoison.Error.t()}

Issues a PATCH request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

patch!(url, body, headers \\ [], options \\ [])

@spec patch!(binary(), any(), headers(), Keyword.t()) ::
  HTTPoison.Response.t()
  | HTTPoison.AsyncResponse.t()
  | HTTPoison.MaybeRedirect.t()

Issues a PATCH request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

persist_schema(dir, package_name)

post(url, body, headers \\ [], options \\ [])

@spec post(binary(), any(), headers(), Keyword.t()) ::
  {:ok,
   HTTPoison.Response.t()
   | HTTPoison.AsyncResponse.t()
   | HTTPoison.MaybeRedirect.t()}
  | {:error, HTTPoison.Error.t()}

Issues a POST request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

post!(url, body, headers \\ [], options \\ [])

@spec post!(binary(), any(), headers(), Keyword.t()) ::
  HTTPoison.Response.t()
  | HTTPoison.AsyncResponse.t()
  | HTTPoison.MaybeRedirect.t()

Issues a POST request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

process_headers(headers)

This function is deprecated. Use process_response_headers/1 instead.

@spec process_headers(list()) :: any()

Callback implementation for HTTPoison.Base.process_headers/1.

process_request_body(body)

@spec process_request_body(body()) :: body()

Callback implementation for HTTPoison.Base.process_request_body/1.

process_request_headers(headers)

@spec process_request_headers(headers()) :: headers()

Callback implementation for HTTPoison.Base.process_request_headers/1.

process_request_options(options)

@spec process_request_options(options()) :: options()

Callback implementation for HTTPoison.Base.process_request_options/1.

process_request_params(params)

@spec process_request_params(params()) :: params()

Callback implementation for HTTPoison.Base.process_request_params/1.

process_request_url(url)

@spec process_request_url(url()) :: url()

Callback implementation for HTTPoison.Base.process_request_url/1.

process_response(response)

@spec process_response(HTTPoison.Base.response()) :: any()

Callback implementation for HTTPoison.Base.process_response/1.

process_response_body(body)

@spec process_response_body(binary()) :: any()

Callback implementation for HTTPoison.Base.process_response_body/1.

process_response_chunk(chunk)

@spec process_response_chunk(binary()) :: any()

Callback implementation for HTTPoison.Base.process_response_chunk/1.

process_response_headers(headers)

@spec process_response_headers(list()) :: any()

Callback implementation for HTTPoison.Base.process_response_headers/1.

process_response_status_code(status_code)

@spec process_response_status_code(integer()) :: any()

Callback implementation for HTTPoison.Base.process_response_status_code/1.

process_status_code(status_code)

This function is deprecated. Use process_response_status_code/1 instead.

@spec process_status_code(integer()) :: any()

Callback implementation for HTTPoison.Base.process_status_code/1.

process_url(url)

This function is deprecated. Use process_request_url/1 instead.

@spec process_url(url()) :: url()

Callback implementation for HTTPoison.Base.process_url/1.

put(url, body \\ "", headers \\ [], options \\ [])

@spec put(binary(), any(), headers(), Keyword.t()) ::
  {:ok,
   HTTPoison.Response.t()
   | HTTPoison.AsyncResponse.t()
   | HTTPoison.MaybeRedirect.t()}
  | {:error, HTTPoison.Error.t()}

Issues a PUT request to the given url.

Returns {:ok, response} if the request is successful, {:error, reason} otherwise.

See request/5 for more detailed information.

put!(url, body \\ "", headers \\ [], options \\ [])

@spec put!(binary(), any(), headers(), Keyword.t()) ::
  HTTPoison.Response.t()
  | HTTPoison.AsyncResponse.t()
  | HTTPoison.MaybeRedirect.t()

Issues a PUT request to the given url, raising an exception in case of failure.

If the request does not fail, the response is returned.

See request!/5 for more detailed information.

request(request)

@spec request(HTTPoison.Request.t()) ::
  {:ok,
   HTTPoison.Response.t()
   | HTTPoison.AsyncResponse.t()
   | HTTPoison.MaybeRedirect.t()}
  | {:error, HTTPoison.Error.t()}

Issues an HTTP request using an HTTPoison.Request struct.

This function returns {:ok, response}, {:ok, async_response}, or {:ok, maybe_redirect} if the request is successful, {:error, reason} otherwise.

Redirect handling

If the option :follow_redirect is given, HTTP redirects are automatically follow if the method is set to :get or :head and the response's status_code is 301, 302 or 307.

If the method is set to :post, then the only status_code that get's automatically followed is 303.

If any other method or status_code is returned, then this function returns a returns a {:ok, %HTTPoison.MaybeRedirect{}} containing the redirect_url for you to re-request with the method set to :get.

Examples

request = %HTTPoison.Request{
  method: :post,
  url: "https://my.website.com",
  body: "{\"foo\": 3}",
  headers: [{"Accept", "application/json"}]
}

request(request)

request(method, url, body \\ "", headers \\ [], options \\ [])

@spec request(method(), binary(), any(), headers(), Keyword.t()) ::
  {:ok,
   HTTPoison.Response.t()
   | HTTPoison.AsyncResponse.t()
   | HTTPoison.MaybeRedirect.t()}
  | {:error, HTTPoison.Error.t()}

Issues an HTTP request with the given method to the given url.

This function is usually used indirectly by get/3, post/4, put/4, etc

Args:

method - HTTP method as an atom (:get, :head, :post, :put, :delete, etc.)

url - target url as a binary string or char list

body - request body. See more below

headers - HTTP headers as an orddict (e.g., [{"Accept", "application/json"}])

options - Keyword list of options

Body: see type HTTPoison.Request

Options: see type HTTPoison.Request

This function returns {:ok, response}, {:ok, async_response}, or {:ok, maybe_redirect} if the request is successful, {:error, reason} otherwise.

Redirect handling

If the option :follow_redirect is given, HTTP redirects are automatically follow if the method is set to :get or :head and the response's status_code is 301, 302 or 307.

If the method is set to :post, then the only status_code that get's automatically followed is 303.

If any other method or status_code is returned, then this function returns a returns a {:ok, %HTTPoison.MaybeRedirect{}} containing the redirect_url for you to re-request with the method set to :get.

Examples

request(:post, "https://my.website.com", "{\"foo\": 3}", [{"Accept", "application/json"}])

request!(request)

@spec request!(HTTPoison.Request.t()) ::
  HTTPoison.Response.t()
  | HTTPoison.AsyncResponse.t()
  | HTTPoison.MaybeRedirect.t()

Issues an HTTP request an HTTPoison.Request struct. exception in case of failure.

request!/1 works exactly like request/1 but it returns just the response in case of a successful request, raising an exception in case the request fails.

request!(method, url, body \\ "", headers \\ [], options \\ [])

@spec request!(method(), binary(), any(), headers(), Keyword.t()) ::
  HTTPoison.Response.t()
  | HTTPoison.AsyncResponse.t()
  | HTTPoison.MaybeRedirect.t()

Issues an HTTP request with the given method to the given url, raising an exception in case of failure.

request!/5 works exactly like request/5 but it returns just the response in case of a successful request, raising an exception in case the request fails.

run(args)

@spec run(any()) :: any()

Callback implementation for Mix.Task.run/1.

start()

Starts HTTPoison and its dependencies.

stream_next(resp)

@spec stream_next(HTTPoison.AsyncResponse.t()) ::
  {:ok, HTTPoison.AsyncResponse.t()} | {:error, HTTPoison.Error.t()}

Requests the next message to be streamed for a given HTTPoison.AsyncResponse.

See request!/5 for more detailed information.

write_schema(dir, package_name, data)