Lightning.Accounts.UserNotifier (Lightning v0.10.0-pre3)

The UserNotifier module.

Functions

build_digest_url(workflow, start_date, end_date)

deliver_confirmation_instructions(user, url)

Deliver instructions to confirm account.

deliver_confirmation_instructions(enroller, user, url)

Deliver instructions to confirm account.

deliver_project_addition_notification(user, project)

Deliver email to notify user of his addition of a project.

deliver_project_digest(digest_data, params)

Deliver a project digest of daily/weekly or monthly activity to a user.

deliver_reset_password_instructions(user, url)

Deliver instructions to reset a user password.

deliver_update_email_instructions(email, url)

Deliver instructions to update a user email.

deliver_update_email_warning(email, new_email)

Deliver warning to update a user email.

notify_project_deletion(user, project)

send_credential_deletion_notification_email(user, credential)

send_deletion_notification_email(user)

Deliver an email to notify the user about their account being deleted