Lightning.Accounts.UserNotifier (Lightning v2.15.13-pre)

The UserNotifier module.

Functions

build_digest_url(workflow, start_date, end_date)

deliver_confirmation_instructions(user, token)

Deliver instructions to confirm account.

deliver_confirmation_instructions(enroller, user, token)

Deliver instructions to confirm account.

deliver_credential_transfer_confirmation_instructions(owner, receiver, credential, token)

Deliver instructions to confirm a credential transfer.

deliver_credential_transfer_notification(receiver, owner, credential)

deliver_project_addition_notification(user, project)

Deliver email to notify user of his addition of a project.

deliver_project_digest(digest_data, params)

Deliver a project digest of daily/weekly or monthly activity to a user.

deliver_project_invitation_email(user, inviter, project, role, token)

deliver_reset_password_instructions(user, url)

Deliver instructions to reset a user password.

deliver_update_email_instructions(user, token)

Deliver instructions to update a user.

deliver_update_email_warning(user, new_email)

Deliver warning to update a user.

notify_history_export_completion(user, project_file)

notify_project_deletion(user, project)

remind_account_confirmation(user, token)

send_credential_deletion_notification_email(user, credential)

send_data_retention_change_email(user, updated_project)

@spec send_data_retention_change_email(
  user :: map(),
  updated_project :: map()
) :: {:ok, term()} | {:error, term()}

Deliver an email to notify the user about a data retention setting change made in their project

send_deletion_notification_email(user)

Deliver an email to notify the user about their account being deleted

send_trigger_failure_mail(user, workflow, timestamp)