Lightning.AiAssistant (Lightning v2.7.1)

The AI assistant module.

Functions

authorized?(user)

@spec authorized?(Lightning.Accounts.User.t()) :: boolean()

Checks if the user is authorized to access the AI assistant.

endpoint_available?()

@spec endpoint_available?() :: boolean()

Checks if the Apollo endpoint is available.

new_session(job)

@spec new_session(Lightning.Workflows.Job.t()) :: Lightning.AiAssistant.Session.t()

Creates a new session with the given job.

Example

iex> AiAssistant.new_session(%Lightning.Workflows.Job{
...>   body: "fn()",
...>   adaptor: "@openfn/language-common@latest"
...> })
%Lightning.AiAssistant.Session{
  expression: "fn()",
  adaptor: "@openfn/language-common@1.6.2",
  history: []
}

ℹ️ The adaptor field is resolved to the latest version when @latest is provided as Apollo expects a specific version.

push_history(session, message)

@spec push_history(Lightning.AiAssistant.Session.t(), %{required(String.t()) => any()}) ::
  Lightning.AiAssistant.Session.t()

query(session, content)

@spec query(Lightning.AiAssistant.Session.t(), String.t()) ::
  {:ok, Lightning.AiAssistant.Session.t()} | :error

Queries the AI assistant with the given content.

Returns {:ok, session} if the query was successful, otherwise :error.

Example

iex> AiAssistant.query(session, "fn()")
{:ok, session}