Lightning.AttemptService (Lightning v0.10.0-pre13)

The Attempts context.

Functions

append(attempt, run)

@spec append(
  Lightning.Attempt.t() | Lightning.AttemptRun.t(),
  Ecto.Changeset.t(Lightning.Invocation.Run.t())
) ::
  {:ok, Lightning.AttemptRun.t()}
  | {:error, Ecto.Changeset.t(Lightning.AttemptRun.t())}

Adds an Attempt to an unsaved Run

When given an Attempt, it simply adds the Run to a new AttemptRun. However when given an AttemptRun, the Run (from the AttemptRun) is set as the previous Run for the new unsaved Run.

build_attempt(work_order, reason)

calculate_runs(workflow, existing_runs, starting_run)

create_attempt(work_order, job, reason)

Create an attempt

Examples

iex> create_attempt(%{field: value})
{:ok, %Attempt{}}

iex> create_attempt(%{field: bad_value})
{:error, %Ecto.Changeset{}}

get_for_rerun(attempt_id, run_id)

get_last_attempt_for(run)

Get the latest attempt associated to a given run

get_workflow_for(attempt)

list_for_rerun_from_job(order_ids, job_id)

@spec list_for_rerun_from_job(
  workorders :: [Ecto.UUID.t(), ...],
  job_id :: Ecto.UUID.t()
) :: [Lightning.AttemptRun.t(), ...]

Returns a list of AttemptRun structs that should be rerun for the given list of workorder ids that are associated to the given Job

list_for_rerun_from_start(order_ids)

@spec list_for_rerun_from_start([Ecto.UUID.t()]) :: [Lightning.AttemptRun.t()]

Returns a list of AttemptRun structs that should be rerun for the given list of work order ids.

retry(attempt, run, reason)

@spec retry(
  Lightning.Attempt.t(),
  Lightning.Invocation.Run.t(),
  Ecto.Changeset.t(Lightning.InvocationReason.t())
  | Lightning.InvocationReason.t()
) :: Ecto.Multi.t()

Creates a new Attempt starting from a given run.

All upstream/prior Runs that were performed on that attempt are associated with the new Attempt, where as the specified run is used to create a new one and is added to the Attempt.

Any runs downstream from the Run given are ignored.

retry_many(attempt_runs, reasons)

@spec retry_many([Lightning.AttemptRun.t()], [Lightning.InvocationReason.t()]) ::
  Ecto.Multi.t()

Creates new Attempts for each pair of corresponding AttemptRun and InvocationReason.