Lightning.AuthProviders.Common (Lightning v2.15.7)

Provides common functionality for handling OAuth authentication across different providers.

Functions

authorize_url(client, state, scopes, opts \\ [])

Constructs the authorization URL with the given client, state, scopes, and options.

build_client(provider, wellknown_url, opts \\ [])

Builds a new OAuth client with the specified configuration, authorization URL, token URL, and options.

get_token(client, params)

Requests an authentication token from the OAuth provider.

get_userinfo(client, token, wellknown_url)

Retrieves user information from the OAuth provider.

get_wellknown(wellknown_url)

Fetches the well-known configuration from the OAuth provider.

get_wellknown!(wellknown_url)

Fetches the well-known configuration from the OAuth provider and raises an error if not successful.

introspect(result, provider, wellknown_url)

refresh_token(client, token)

Refreshes the authentication token using the OAuth provider.

still_fresh(token_body, threshold \\ 5, time_unit \\ :minute)

Checks if a token is still valid or must be refreshed. If expires_at is nil, it will return false, forcing a refresh. If the token has already expired or will expire before the default buffer (in the next 5 minutes) we return false, forcing a refresh.