Lightning.Config behaviour (Lightning v2.4.5)

Centralised runtime configuration for Lightning.

Callbacks

grace_period()

@callback grace_period() :: integer()

purge_deleted_after_days()

@callback purge_deleted_after_days() :: integer()

repo_connection_token_signer()

@callback repo_connection_token_signer() :: Joken.Signer.t()

run_token_signer()

@callback run_token_signer() :: Joken.Signer.t()

worker_secret()

@callback worker_secret() :: binary() | nil

worker_token_signer()

@callback worker_token_signer() :: Joken.Signer.t()

Functions

grace_period()

purge_deleted_after_days()

repo_connection_token_signer()

run_token_signer()

Returns the Token signer used to sign and verify run tokens.

worker_secret()

worker_token_signer()

Returns the Token signer used to verify worker tokens.