Lightning.Config behaviour (Lightning v2.15.11-pre)

Centralised runtime configuration for Lightning.

Callbacks

activity_cleanup_chunk_size()

@callback activity_cleanup_chunk_size() :: integer()

adaptor_registry()

@callback adaptor_registry() :: Keyword.t()

ai_assistant_modes()

@callback ai_assistant_modes() :: %{required(atom()) => module()}

apollo(key)

@callback apollo(key :: atom() | nil) :: map()

book_demo_banner_enabled?()

@callback book_demo_banner_enabled?() :: boolean()

book_demo_calendly_url()

@callback book_demo_calendly_url() :: String.t()

book_demo_openfn_workflow_url()

@callback book_demo_openfn_workflow_url() :: String.t()

broadcast_work_available?()

@callback broadcast_work_available?() :: boolean()

check_flag?(atom)

@callback check_flag?(atom()) :: boolean() | nil

cors_origin()

@callback cors_origin() :: list()

credential_transfer_token_validity_in_days()

@callback credential_transfer_token_validity_in_days() :: integer()

default_ecto_database_timeout()

@callback default_ecto_database_timeout() :: integer()

default_max_run_duration()

@callback default_max_run_duration() :: integer()

email_sender_name()

@callback email_sender_name() :: String.t()

env()

@callback env() :: :dev | :test | :prod

external_metrics_module()

@callback external_metrics_module() :: module() | nil

gdpr_banner()

@callback gdpr_banner() :: map() | false

gdpr_preferences()

@callback gdpr_preferences() :: map() | false

get_extension_mod(key)

@callback get_extension_mod(key :: atom()) :: any()

google(key)

@callback google(key :: atom()) :: any()

grace_period()

@callback grace_period() :: integer()

instance_admin_email()

@callback instance_admin_email() :: String.t()

kafka_alternate_storage_enabled?()

@callback kafka_alternate_storage_enabled?() :: boolean()

kafka_alternate_storage_file_path()

@callback kafka_alternate_storage_file_path() :: String.t()

kafka_duplicate_tracking_retention_seconds()

@callback kafka_duplicate_tracking_retention_seconds() :: integer()

kafka_notification_embargo_seconds()

@callback kafka_notification_embargo_seconds() :: integer()

kafka_number_of_consumers()

@callback kafka_number_of_consumers() :: integer()

kafka_number_of_messages_per_second()

@callback kafka_number_of_messages_per_second() :: float()

kafka_number_of_processors()

@callback kafka_number_of_processors() :: integer()

kafka_triggers_enabled?()

@callback kafka_triggers_enabled?() :: boolean()

max_credential_sensitive_values()

@callback max_credential_sensitive_values() :: pos_integer()

metrics_run_performance_age_seconds()

@callback metrics_run_performance_age_seconds() :: integer()

metrics_run_queue_metrics_period_seconds()

@callback metrics_run_queue_metrics_period_seconds() :: integer()

metrics_stalled_run_threshold_seconds()

@callback metrics_stalled_run_threshold_seconds() :: integer()

metrics_unclaimed_run_threshold_seconds()

@callback metrics_unclaimed_run_threshold_seconds() :: integer()

oauth_provider(key)

@callback oauth_provider(key :: atom()) :: keyword() | nil

per_workflow_claim_limit()

@callback per_workflow_claim_limit() :: pos_integer()

promex_expensive_metrics_enabled?()

@callback promex_expensive_metrics_enabled?() :: boolean()

promex_metrics_endpoint_authorization_required?()

@callback promex_metrics_endpoint_authorization_required?() :: boolean()

promex_metrics_endpoint_scheme()

@callback promex_metrics_endpoint_scheme() :: String.t()

promex_metrics_endpoint_token()

@callback promex_metrics_endpoint_token() :: String.t()

purge_deleted_after_days()

@callback purge_deleted_after_days() :: integer()

repo_connection_token_signer()

@callback repo_connection_token_signer() :: Joken.Signer.t()

reset_password_token_validity_in_days()

@callback reset_password_token_validity_in_days() :: integer()

run_token_signer()

@callback run_token_signer() :: Joken.Signer.t()

runtime_manager_port()

@callback runtime_manager_port() :: integer()

sentry()

@callback sentry() :: module()

storage()

@callback storage() :: term()

storage(key)

@callback storage(key :: atom()) :: term()

token_signer()

@callback token_signer() :: Joken.Signer.t()

ui_metrics_tracking_enabled?()

@callback ui_metrics_tracking_enabled?() :: boolean()

usage_tracking()

@callback usage_tracking() :: Keyword.t()

usage_tracking_cleartext_uuids_enabled?()

@callback usage_tracking_cleartext_uuids_enabled?() :: boolean()

usage_tracking_cron_opts()

@callback usage_tracking_cron_opts() :: [Oban.Plugins.Cron.cron_input()]

usage_tracking_enabled?()

@callback usage_tracking_enabled?() :: boolean()

usage_tracking_host()

@callback usage_tracking_host() :: String.t()

usage_tracking_run_chunk_size()

@callback usage_tracking_run_chunk_size() :: integer()

webhook_response_timeout_ms()

@callback webhook_response_timeout_ms() :: integer()

webhook_retry()

@callback webhook_retry() :: Keyword.t()

webhook_retry(key)

@callback webhook_retry(key :: atom()) :: any()

worker_secret()

@callback worker_secret() :: binary() | nil

worker_token_signer()

@callback worker_token_signer() :: Joken.Signer.t()

Functions

activity_cleanup_chunk_size()

adaptor_registry()

Returns the configuration for the Lightning.AdaptorRegistry service

ai_assistant_modes()

apollo(key \\ nil)

Returns the Apollo server configuration.

book_demo_banner_enabled?()

book_demo_calendly_url()

book_demo_openfn_workflow_url()

broadcast_work_available?()

check_flag?(flag)

cors_origin()

credential_transfer_token_validity_in_days()

default_ecto_database_timeout()

default_max_run_duration()

Returns the default maximum run duration in seconds.

email_sender_name()

env()

Returns the environment.

Either :dev, :test, or :prod.

external_metrics_module()

gdpr_banner()

gdpr_preferences()

get_extension_mod(key)

google(key)

grace_period()

The grace period is configurable and is used to wait for an additional amount of time after a given run was meant to be finished.

The returned value is in seconds.

instance_admin_email()

kafka_alternate_storage_enabled?()

kafka_alternate_storage_file_path()

kafka_duplicate_tracking_retention_seconds()

kafka_notification_embargo_seconds()

kafka_number_of_consumers()

kafka_number_of_messages_per_second()

kafka_number_of_processors()

kafka_triggers_enabled?()

max_credential_sensitive_values()

Returns the maximum number of sensitive values allowed in a credential body.

This limit prevents performance issues with the scrubber when credentials have very large bodies with many secrets.

metrics_run_performance_age_seconds()

metrics_run_queue_metrics_period_seconds()

metrics_stalled_run_threshold_seconds()

metrics_unclaimed_run_threshold_seconds()

oauth_provider(key)

per_workflow_claim_limit()

promex_expensive_metrics_enabled?()

promex_metrics_endpoint_authorization_required?()

promex_metrics_endpoint_scheme()

promex_metrics_endpoint_token()

purge_deleted_after_days()

repo_connection_token_signer()

reset_password_token_validity_in_days()

run_token_signer()

Returns the Token signer used to sign and verify run tokens.

runtime_manager_port()

sentry()

storage()

storage(key)

token_signer()

ui_metrics_tracking_enabled?()

usage_tracking()

usage_tracking_cleartext_uuids_enabled?()

usage_tracking_cron_opts()

usage_tracking_enabled?()

usage_tracking_host()

usage_tracking_run_chunk_size()

webhook_response_timeout_ms()

webhook_retry()

webhook_retry(key)

worker_secret()

worker_token_signer()

Returns the Token signer used to verify worker tokens.