Lightning.Credentials (Lightning v0.3.1)

The Credentials context.

Functions

change_credential(credential, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking credential changes.

Examples

iex> change_credential(credential)
%Ecto.Changeset{data: %Credential{}}

create_credential(attrs \\ %{})

Creates a credential.

Examples

iex> create_credential(%{field: value})
{:ok, %Credential{}}

iex> create_credential(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_credential(credential)

Deletes a credential.

Examples

iex> delete_credential(credential)
{:ok, %Credential{}}

iex> delete_credential(credential)
{:error, %Ecto.Changeset{}}

get_credential!(id)

Gets a single credential.

Raises Ecto.NoResultsError if the Credential does not exist.

Examples

iex> get_credential!(123)
%Credential{}

iex> get_credential!(456)
** (Ecto.NoResultsError)

invalid_projects_for_user(credential_id, user_id)

Given a credential and a user, returns a list of invalid projects—i.e., those that the credential is shared with but that the user does not have access to.

This is used to generate a validation error when a credential cannot be transferred.

Examples

iex> can_credential_be_shared_to_user(credential_id, user_id)
[]

iex> can_credential_be_shared_to_user(credential_id, user_id)
["52ea8758-6ce5-43d7-912f-6a1e1f11dc55"]

list_credentials()

Returns the list of credentials.

Examples

iex> list_credentials()
[%Credential{}, ...]

list_credentials(project)

list_credentials_for_user(user_id)

Returns the list of credentials for a given user.

Examples

iex> list_credentials_for_user(123)
[%Credential{user_id: 123}, %Credential{user_id: 123},...]

sensitive_values_for(id)

@spec sensitive_values_for(Ecto.UUID.t() | Lightning.Credentials.Credential.t() | nil) ::
  [any()]

update_credential(credential, attrs)

Updates a credential.

Examples

iex> update_credential(credential, %{field: new_value})
{:ok, %Credential{}}

iex> update_credential(credential, %{field: bad_value})
{:error, %Ecto.Changeset{}}