Lightning.Credentials (Lightning v2.4.13)

The Credentials context.

Functions

basic_auth_for(credential)

cancel_scheduled_deletion(credential_id)

change_credential(credential, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking credential changes.

Examples

iex> change_credential(credential)
%Ecto.Changeset{data: %Credential{}}

create_credential(attrs \\ %{})

Creates a credential.

Examples

iex> create_credential(%{field: value})
{:ok, %Credential{}}

iex> create_credential(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_credential(credential)

Deletes a credential.

Examples

iex> delete_credential(credential)
{:ok, %Credential{}}

iex> delete_credential(credential)
{:error, %Ecto.Changeset{}}

get_credential!(id)

Gets a single credential.

Raises Ecto.NoResultsError if the Credential does not exist.

Examples

iex> get_credential!(123)
%Credential{}

iex> get_credential!(456)
** (Ecto.NoResultsError)

get_credential_by_project_credential(project_credential_id)

get_credential_for_update!(id)

get_schema(schema_name)

@spec get_schema(String.t()) :: Lightning.Credentials.Schema.t()

Creates a credential schema from credential json schema.

has_activity_in_projects?(credential)

Checks if a given Credential has any associated Step activity.

Parameters

_credential: A Credential struct. Only the id field is used by the function.

Returns

true if there's at least one Step associated with the given Credential.

false otherwise.

Examples

iex> has_activity_in_projects?(%Credential{id: some_id})
true

iex> has_activity_in_projects?(%Credential{id: another_id})
false

Notes

This function leverages the association between Step and Credential to determine if any steps exist for a given credential. It's a fast check that does not load any records into memory, but simply checks for their existence.

invalid_projects_for_user(credential_id, user_id)

Given a credential and a user, returns a list of invalid projects—i.e., those that the credential is shared with but that the user does not have access to.

This is used to generate a validation error when a credential cannot be transferred.

Examples

iex> can_credential_be_shared_to_user(credential_id, user_id)
[]

iex> can_credential_be_shared_to_user(credential_id, user_id)
["52ea8758-6ce5-43d7-912f-6a1e1f11dc55"]

list_credentials()

Returns the list of credentials.

Examples

iex> list_credentials()
[%Credential{}, ...]

list_credentials(project)

list_credentials_for_user(user_id)

Returns the list of credentials for a given user.

Examples

iex> list_credentials_for_user(123)
[%Credential{user_id: 123}, %Credential{user_id: 123},...]

lookup_adapter(schema)

maybe_refresh_token(credential)

migrate_credential_body(credential)

perform(job)

Perform, when called with %{"type" => "purge_deleted"} will find credentials that are ready for permanent deletion, set their bodies to null, and try to purge them.

schedule_credential_deletion(credential)

Schedules a given credential for deletion.

The deletion date is determined based on the :purge_deleted_after_days configuration in the application environment. If this configuration is absent, the credential is scheduled for immediate deletion.

The function will also perform necessary side effects such as:

Removing associations of the credential.

Notifying the owner of the credential about the scheduled deletion.

Parameters

credential: A Credential struct that is to be scheduled for deletion.

Returns

{:ok, credential}: Returns an :ok tuple with the updated credential struct if the update was successful.

{:error, changeset}: Returns an :error tuple with the changeset if the update failed.

Examples

iex> schedule_credential_deletion(%Credential{id: some_id})
{:ok, %Credential{}}

iex> schedule_credential_deletion(%Credential{})
{:error, %Ecto.Changeset{}}

sensitive_values_for(id)

@spec sensitive_values_for(Ecto.UUID.t() | Lightning.Credentials.Credential.t() | nil) ::
  [any()]

update_credential(credential, attrs)

Updates a credential.

Examples

iex> update_credential(credential, %{field: new_value})
{:ok, %Credential{}}

iex> update_credential(credential, %{field: bad_value})
{:error, %Ecto.Changeset{}}