Lightning.Invocation.LogLine (Lightning v2.15.11-pre)

Ecto model for run logs.

Every log line from a worker is stored using this model.

Currently the :message field is a text field, however the worker sends messages in a JSON format. We will at some point consider changing this to a JSON field; but for now we coerce it into a string.

See LogMessage for more information.

Types

t()

@type t() :: %Lightning.Invocation.LogLine{
  __meta__: Ecto.Schema.Metadata.t(),
  id: Ecto.UUID.t() | nil,
  level: term(),
  message: String.t(),
  run: Lightning.Run.t() | Ecto.Association.NotLoaded.t() | nil,
  run_id: term(),
  source: term(),
  step: Lightning.Invocation.Step.t() | Ecto.Association.NotLoaded.t() | nil,
  step_id: term(),
  timestamp: DateTime.t()
}

Functions

new(run, attrs \\ %{}, scrubber)

validate(changeset, scrubber \\ nil)