Lightning.Invocation.Query (Lightning v2.5.0)

Query functions for working with Steps and Dataclips

Functions

any_step()

To be used in preloads for workflow > job > step when the presence of any step is all the information we need. As in, "Does this job have any steps?"

dataclip_with_body()

By default, the dataclip body is not returned via a query. This query selects the body specifically.

last_n_for_job(job_id, limit)

last_step_for_job(job)

@spec last_step_for_job(Lightning.Workflows.Job.t()) :: Ecto.Queryable.t()

The last step for a job

last_successful_step_for_job(job)

@spec last_successful_step_for_job(Lightning.Workflows.Job.t()) :: Ecto.Queryable.t()

The last successful step for a job, used in scheduler to enable downstream runs to access a previous run's state

select_as_input(query)

Returns a dataclip formatted for use as an input state.

Only http_request dataclips are changed, their body is nested inside a "data" key and request data is added as a "request" key.

steps_for(user)

@spec steps_for(Lightning.Accounts.User.t()) :: Ecto.Queryable.t()

Steps for a specific user

steps_with_reason(query, exit_reason)

@spec steps_with_reason(Ecto.Queryable.t(), String.t()) :: Ecto.Queryable.t()

The last step for a job for a particular exit reason, used in scheduler

wipe_dataclips(query \\ Dataclip)