Lightning.Jobs.Query (Lightning v0.1.11)

Query module for finding Jobs.

Functions

downstream_jobs_query(upstream_query)

Find the 'descendants' for a Job in a workflow, based on a given Job. This query uses a recursive CTE to interate downwards following the upstream job relationship until it can't find any more jobs.

enabled_cron_jobs()

@spec enabled_cron_jobs() :: Ecto.Queryable.t()

Returns active jobs with their cron triggers for use in the cron scheduling service.

initiating_jobs_query(project)

Find all the initiating jobs for a project, the criteria is that the job is not downstream from another job.

jobs_for(user)

@spec jobs_for(Lightning.Accounts.User.t()) :: Ecto.Queryable.t()

Returns all jobs accessible to a user, via their projects

upstream_job_query(job)

Find the 'first' Job in a workflow, based on a given Job. This query uses a recursive CTE to interate upwards until it finds the first Job that has no upstream job.

workflow_query(job)

Find all Jobs related to either a Job or a Project.

When given a Job, we first find the initiating/top job and then select all descendants - ensuring that you get sibling jobs as well (two or more jobs that rely on the same upstream job).

When given a Project, we can find all initiating jobs by their lack of an upstream job, then find all descendants. In addition a workflow_id is attached to the query in order to later group workflows by their initiating job.