Lightning.KafkaTriggers (Lightning v2.15.0-pre4)

Contains the logic to manage kafka trigger and their associated pipelines.

Functions

alternate_storage_file_name(trigger_id, message)

build_topic_partition_offset(message)

Generate the key that is used to identify duplicate messages when used in association with the trigger id.

convert_rate_limit()

enable_disable_triggers(triggers)

failure_notification_tracking_key(trigger_id)

find_enabled_triggers()

generate_pipeline_child_spec(trigger)

Generate the child spec needed to start a Pipeline child process.

get_kafka_triggers_being_updated(changeset)

initial_policy(map)

maybe_write_to_alternate_storage(trigger_id, msg)

notify_users_of_trigger_failure(trigger_id)

send_notification?(sending_at, last_sent_at)

start_triggers()

update_pipeline(supervisor, trigger_id)