Lightning.Pipeline.StateAssembler (Lightning v0.5.0)

Facilities for building the state for a Run

How state is assembled

For the most common jobs, an inbound webhook will store an :http_request type dataclip. The reason that is created is associated with the dataclip.

At runtime, the initial state for a Run will be in the shape of:

{ "data": <the dataclip>, "configuration": <the job's credential> }

Saved inputs

Saved custom inputs will only have state.configuration changed, everything else will remain as displayed.

Flow Jobs

When a Job is triggered by a previous Jobs success or failure these are the rules for constructing that Jobs state:

For jobs that trigger on it's upstream jobs failure, the event will have the previous runs input dataclip as its input dataclip.

The state will also have the log of the previous run attached on the error key.

For Jobs triggered by a previous success, the run will have the previous runs output dataclip as its input dataclip.

:run_result dataclips are expected to already have a data key, and are merged into the root.

Functions

assemble(run)

@spec assemble(run :: Lightning.Invocation.Run.t()) :: String.t()

Assemble state for use in a Run.