Lightning.Projects (Lightning v0.10.0-pre4)

The Projects context.

Functions

cancel_scheduled_deletion(project_id)

change_project(project, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking project changes.

Examples

iex> change_project(project)
%Ecto.Changeset{data: %Project{}}

create_project(attrs \\ %{})

Creates a project.

Examples

iex> create_project(%{field: value})
{:ok, %Project{}}

iex> create_project(%{field: bad_value})
{:error, %Ecto.Changeset{}}

delete_project(project)

Deletes a project and its related data, including workflows, work orders, runs, jobs, attempts, triggers, project users, project credentials, and dataclips

Examples

iex> delete_project(project)
{:ok, %Project{}}

iex> delete_project(project)
{:error, %Ecto.Changeset{}}

export_project(atom, project_id)

@spec export_project(:yaml, any()) :: {:ok, binary()}

Exports a project as yaml.

Examples

iex> export_project(:yaml, project_id)
{:ok, string}

get_project(id)

get_project!(id)

Gets a single project.

Raises Ecto.NoResultsError if the Project does not exist.

Examples

iex> get_project!(123)
%Project{}

iex> get_project!(456)
** (Ecto.NoResultsError)

get_project_credential(project_id, credential_id)

get_project_user(id)

get_project_user(project, user)

get_project_user!(id)

Gets a single project_user.

Raises Ecto.NoResultsError if the ProjectUser does not exist.

Examples

iex> get_project_user!(123)
%ProjectUser{}

iex> get_project_user!(456)
** (Ecto.NoResultsError)

get_project_user_role(user, project)

Returns the role of a user in a project. Possible roles are :admin, :viewer, :editor, and :owner

Examples

iex> get_project_user_role(user, project)
:admin

iex> get_project_user_role(user, project)
:viewer

iex> get_project_user_role(user, project)
:editor

iex> get_project_user_role(user, project)
:owner

get_project_with_users!(id)

Gets a single project with it's members via project_users.

Raises Ecto.NoResultsError if the Project does not exist.

Examples

iex> get_project!(123)
%Project{}

iex> get_project!(456)
** (Ecto.NoResultsError)

get_projects_for_user(user)

@spec get_projects_for_user(user :: Lightning.Accounts.User.t()) :: [
  Lightning.Projects.Project.t()
]

is_member_of?(project, user)

list_project_credentials(project)

@spec list_project_credentials(project :: Lightning.Projects.Project.t()) :: [
  Lightning.Projects.ProjectCredential.t()
]

list_projects()

Returns the list of projects.

Examples

iex> list_projects()
[%Project{}, ...]

perform(job)

Perform, when called with %{"type" => "purge_deleted"} will find projects that are ready for permanent deletion and purge them.

project_attempt_run_query(project)

project_attempts_query(project)

project_credentials_query(project)

project_dataclip_invocation_reason(project)

project_dataclips_query(project)

project_jobs_query(project)

project_run_invocation_reasons(project)

project_runs_query(project)

project_trigger_invocation_reason(project)

project_triggers_query(project)

project_user_role_query(user, project)

@spec project_user_role_query(
  user :: Lightning.Accounts.User.t(),
  project :: Lightning.Projects.Project.t()
) :: Ecto.Queryable.t()

project_users_query(project)

@spec project_users_query(atom() | %{:id => any(), optional(any()) => any()}) ::
  Ecto.Query.t()

project_workflows_query(project)

project_workorders_query(project)

projects_for_user_query(user)

@spec projects_for_user_query(user :: Lightning.Accounts.User.t()) ::
  Ecto.Queryable.t()

schedule_project_deletion(project)

Given a project, this function sets a scheduled deletion date based on the PURGE_DELETED_AFTER_DAYS environment variable. If no ENV is set, this date defaults to NOW but the automatic project purge cronjob will never run. (Note that subsequent logins will be blocked for projects pending deletion.)

select_first_project_for_user(user)

@spec select_first_project_for_user(user :: Lightning.Accounts.User.t()) ::
  Lightning.Projects.Project.t() | nil

update_project(project, attrs)

Updates a project.

Examples

iex> update_project(project, %{field: new_value})
{:ok, %Project{}}

iex> update_project(project, %{field: bad_value})
{:error, %Ecto.Changeset{}}

update_project_user(project_user, attrs)

Updates a project user.

Examples

iex> update_project_user(project_user, %{field: new_value})
{:ok, %ProjectUser{}}

iex> update_project_user(projectUser, %{field: bad_value})
{:error, %Ecto.Changeset{}}

url_safe_project_name(name)

validate_for_deletion(project, attrs)

Returns an %Ecto.Changeset{} for changing the project scheduled_deletion.

Examples

iex> validate_for_deletion(project)
%Ecto.Changeset{data: %Project{}}