Lightning.Projects.Provisioner (Lightning v2.5.2)

Provides functions for importing projects. This module is used by the provisioning HTTP API.

When providing a project to import, all records must have an id field. It's up to the caller to ensure that the id is unique and generated ahead of time in the case of new records.

Functions

import_document(project, user, data)

@spec import_document(
  Lightning.Projects.Project.t() | nil,
  Lightning.Accounts.User.t()
  | Lightning.VersionControl.ProjectRepoConnection.t(),
  map()
) ::
  {:error,
   Ecto.Changeset.t(Lightning.Projects.Project.t())
   | Lightning.Extensions.UsageLimiting.message()}
  | {:ok, Lightning.Projects.Project.t()}

Import a project.

parse_document(project, data)

@spec parse_document(Lightning.Projects.Project.t(), map()) ::
  Ecto.Changeset.t(Lightning.Projects.Project.t())

preload_dependencies(project)

@spec preload_dependencies(Lightning.Projects.Project.t()) ::
  Lightning.Projects.Project.t()

Preload all dependencies for a project.

Exclude deleted workflows.

validate_extraneous_params(changeset)

Validate that there are no extraneous parameters in the changeset.

For all params in the changeset, ensure that the param is in the list of known fields in the schema.