Lightning.TaskWorker (Lightning v2.15.0-pre5)

A TaskWorker with concurrency limits.

A simple concurrency limiter that wraps Task.Supervisor, which already does have the ability to specify max_children; it throws an error when that limit is exceeded.

To use it, start it like any other process; ideally in your supervision tree.

  ...,
  {Lightning.TaskWorker, name: :cli_task_worker, max_tasks: 4}

Options

:max_tasks Defaults to the number of system schedulers available to the vm.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

get_status(worker)

start_link(opts \\ [name: nil])

start_task(worker, fun)

@spec start_task(worker :: GenServer.name(), (-> any())) ::
  {:error, :too_many_processes} | term()