Lightning.VersionControl (Lightning v0.10.4)

Boundary module for handling Version control activities for project, jobs workflows etc Use this module to create, modify and delete connections as well as running any associated sync jobs

Functions

add_github_installation_id(user_id, installation_id)

add_github_repo_and_branch(project_id, repo, branch)

create_github_connection(attrs)

Creates a connection between a project and a github repo

fetch_installation_repos(project_id)

fetch_repo_branches(project_id, repo_name)

get_pending_user_installation(user_id)

@spec get_pending_user_installation(Ecto.UUID.t()) ::
  Lightning.VersionControl.ProjectRepoConnection.t() | nil

get_repo_connection(project_id)

github_enabled?()

remove_github_connection(project_id)

Deletes a github connection used when re installing

run_sync(project_id, user_name)