Lightning.WorkflowTemplates (Lightning v2.15.14)

The WorkflowTemplates context.

Functions

change_template(template, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking template changes.

Examples

iex> change_template(template)
%Ecto.Changeset{data: %WorkflowTemplate{}}

create_template(attrs \\ %{})

Creates or updates a workflow template.

If a template already exists for the workflow, it will be updated. Otherwise, a new template will be created.

Examples

iex> create_template(%{name: "My Template", code: "workflow code", workflow_id: "123"})
{:ok, %WorkflowTemplate{}}

iex> create_template(%{name: "Invalid"})
{:error, %Ecto.Changeset{}}

delete_template(template)

Deletes a workflow template.

Examples

iex> delete_template(template)
{:ok, %WorkflowTemplate{}}

iex> delete_template(template)
{:error, %Ecto.Changeset{}}

get_template(id)

Gets a single workflow template.

Returns nil if the Workflow template does not exist.

Examples

iex> get_template(123)
%WorkflowTemplate{}

iex> get_template(456)
nil

get_template!(id)

Gets a single workflow template.

Raises Ecto.NoResultsError if the Workflow template does not exist.

Examples

iex> get_template!(123)
%WorkflowTemplate{}

iex> get_template!(456)
** (Ecto.NoResultsError)

get_template_by_workflow_id(workflow_id)

Gets a template by workflow ID.

Examples

iex> get_template_by_workflow_id("123")
%WorkflowTemplate{}

iex> get_template_by_workflow_id("456")
nil

list_templates()

Lists all workflow templates.

Examples

iex> list_templates()
[%WorkflowTemplate{}, ...]

list_workflow_templates(workflow)

Lists workflow templates for a specific workflow.

Examples

iex> list_workflow_templates(workflow)
[%WorkflowTemplate{}, ...]

update_template(template, attrs)

Updates a workflow template.

Examples

iex> update_template(template, %{name: "New Name"})
{:ok, %WorkflowTemplate{}}

iex> update_template(template, %{name: ""})
{:error, %Ecto.Changeset{}}