Lightning.Workflows (Lightning v2.0.0-rc11)

The Workflows context.

Functions

build_trigger(attrs)

Builds a Trigger

change_workflow(workflow, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking workflow changes.

Examples

iex> change_workflow(workflow)
%Ecto.Changeset{data: %Workflow{}}

create_edge(attrs)

Creates an edge

create_workflow(attrs \\ %{})

Creates a workflow.

Examples

iex> create_workflow(%{field: value})
{:ok, %Workflow{}}

iex> create_workflow(%{field: bad_value})
{:error, %Ecto.Changeset{}}

get_edge_by_trigger(trigger)

Gets an Edge by its associated Trigger.

Parameters

%Trigger{id: trigger_id}: A Trigger struct from which the associated Edge is to be found.

Returns

Returns an Edge struct preloaded with its source_trigger and target_job if found.

Returns nil if no Edge is associated with the given Trigger.

Examples

trigger = %Trigger{id: 1, ...}
Lightning.Workflows.get_edge_by_trigger(trigger)
# => %Edge{source_trigger: %Trigger{}, target_job: %Job{}, ...}

non_existent_trigger = %Trigger{id: 999, ...}
Lightning.Workflows.get_edge_by_trigger(non_existent_trigger)
# => nil

get_edges_for_cron_execution(datetime)

@spec get_edges_for_cron_execution(DateTime.t()) :: [Lightning.Workflows.Edge.t()]

Returns a list of edges with jobs to execute, given a current timestamp in Unix. This is used by the scheduler, which calls this function once every minute.

get_trigger_by_webhook(path)

Gets a single Trigger by its custom_path or id.

Parameters

path: A binary string representing the custom_path or id of the trigger.

Returns

Returns a Trigger struct if a trigger is found.

Returns nil if no trigger is found for the given path.

Examples

Lightning.Workflows.get_trigger_by_webhook("some_path_or_id")
# => %Trigger{id: 1, custom_path: "some_path_or_id", ...}

Lightning.Workflows.get_trigger_by_webhook("non_existent_path_or_id")
# => nil

get_webhook_trigger(path, opts \\ [])

Gets a Single Edge by it's webhook trigger.

get_workflow(id)

get_workflow!(id)

Gets a single workflow.

Raises Ecto.NoResultsError if the Workflow does not exist.

Examples

iex> get_workflow!(123)
%Workflow{}

iex> get_workflow!(456)
** (Ecto.NoResultsError)

get_workflows_for(project)

@spec get_workflows_for(Lightning.Projects.Project.t()) :: [
  Lightning.Workflows.Workflow.t()
]

Retrieves a list of active Workflows with their jobs and triggers preloaded.

jobs_ordered_subquery()

A way to ensure the consistency of nodes. This query orders jobs based on their inserted_at timestamps in ascending order

list_workflows()

Returns the list of workflows.

Examples

iex> list_workflows()
[%Workflow{}, ...]

mark_for_deletion(workflow, attrs \\ %{})

Returns an %Ecto.Changeset{} for changing the workflow request_deletion.

Examples

iex> change_request_deletion(workflow)
%Ecto.Changeset{data: %Workflow{}}

to_project_space(workflows)

@spec to_project_space([Lightning.Workflows.Workflow.t()]) :: %{}

update_trigger(trigger, attrs)

Updates a trigger

update_workflow(workflow, attrs)

Updates a workflow.

Examples

iex> update_workflow(workflow, %{field: new_value})
{:ok, %Workflow{}}

iex> update_workflow(workflow, %{field: bad_value})
{:error, %Ecto.Changeset{}}

workflow_exists?(project_id, workflow_name)

Check if workflow exist