LightningWeb.Components.NewInputs (Lightning v2.5.4)

A temporary module that will serve as a place to put new inputs that conform with the newer CoreComponents conventions introduced in Phoenix 1.7.

Functions

button(assigns)

Renders a button.

Examples

<.button>Send!</.button>
<.button phx-click="go" class="ml-2">Send!</.button>

Attributes

id (:string) - Defaults to "".

type (:string) - Defaults to "button".

class (:any) - Defaults to "".

color_class (:any) - Defaults to "bg-primary-600 hover:bg-primary-700 text-white focus:ring-primary-500 disabled:bg-primary-300".

tooltip (:any) - Defaults to nil.

Global attributes are accepted. Supports all globals plus: ["disabled", "form", "name", "value"].

Slots

inner_block (required)

error(assigns)

Generates a generic error message.

Slots

inner_block (required)

icon(assigns)

Renders a Heroicon.

Heroicons come in three styles – outline, solid, and mini. By default, the outline style is used, but solid and mini may be applied by using the -solid and -mini suffix.

You can customize the size and colors of the icons by setting width, height, and background color classes.

Icons are extracted from your assets/vendor/heroicons directory and bundled within your compiled app.css by the plugin in your assets/tailwind.config.js.

Examples

<.icon name="hero-x-mark-solid" />
<.icon name="hero-play-circle" class="ml-1 w-3 h-3 animate-spin" />

Attributes

name (:string) (required)

class (:string) - Defaults to nil.

naked (:boolean) - Defaults to false.

Global attributes are accepted.

input(assigns)

Renders an input with label and error messages.

A Phoenix.HTML.FormField may be passed as argument, which is used to retrieve the input name, id, and values. Otherwise all attributes may be passed explicitly.

Types

This function accepts all HTML input types, considering that:

You may also set type="select" to render a <select> tag

type="checkbox" is used exclusively to render boolean values

For live file uploads, see Phoenix.Component.live_file_input/1

See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input for more information.

Examples

<.input field={@form[:email]} type="email" />
<.input name="my-input" errors={["oh no!"]} />

Attributes

id (:any) - Defaults to nil.

name (:any)

label (:string) - Defaults to nil.

value (:any)

type (:string) - Defaults to "text".

field (Phoenix.HTML.FormField) - a form field struct retrieved from the form, for example: @form[:email].

errors (:list) - Defaults to [].

checked (:boolean) - the checked flag for checkbox inputs.

checked_value (:any) - the value to be sent when the checkbox is checked. Defaults to 'true'. Defaults to "true".

unchecked_value (:any) - the value to be sent when the checkbox is unchecked, Defaults to 'false'. Defaults to "false".

hidden_input (:boolean) - controls if this function will generate a hidden input to submit the unchecked value or not. Defaults to 'true'. Defaults to true.

prompt (:string) - the prompt for select inputs. Defaults to nil.

options (:list) - the options to pass to Phoenix.HTML.Form.options_for_select/2.

multiple (:boolean) - the multiple flag for select inputs. Defaults to false.

button_placement (:string) - Defaults to nil.

class (:string) - Defaults to "".

Global attributes are accepted. Supports all globals plus: ["accept", "autocomplete", "capture", "cols", "disabled", "form", "list", "max", "maxlength", "min", "minlength", "multiple", "pattern", "placeholder", "readonly", "required", "rows", "size", "step"].

Slots

inner_block

label(assigns)

Renders a label.

Attributes

for (:any) - Defaults to nil.

class (:any) - Defaults to "".

Slots

inner_block (required)