LightningWeb.Components.Viewers (Lightning v2.15.0-pre6)

Components for rendering Logs and Dataclips.

Scrolling can be tricky

We seldom know how long a log or a dataclip will be, and we want to be able to contain the element in a fixed height container. In some situations wrapping the component in a div with inline-flex class will help with scrolling.

Functions

DataclipViewer(assigns)

Attributes

id (:string) (required)

dataclipId (:map) (required)

react-id (:string) - Defaults to nil.

react-portal-target (:string) - Defaults to nil.

dataclip_viewer(assigns)

log_viewer(assigns)

Renders out a log line stream

Internally it uses the LogLineHighlight hook to highlight the log line with the highlight_id attribute.

Example

<Viewers.log_viewer
  id="log-viewer-data"
  stream={@log_lines}
  highlight_id={@selected_step_id}
/>

Attributes

id (:string) (required)

run_id (:string) (required)

run_state (:any) (required)

logs_empty? (:boolean) (required)

selected_step_id (:string)

current_user (Lightning.Accounts.User) - for checking log filter preference. Defaults to nil.

class (:string) - Additional classes to add to the log viewer container. Defaults to nil.

step_dataclip_viewer(assigns)

Attributes

id (:string) (required)

run_state (:any) (required)

class (:string) - Additional classes to add to the log viewer container. Defaults to nil.

step (:map)

dataclip (Lightning.Invocation.Dataclip)

input_or_output (:atom) (required) - Must be one of :input, or :output.

project_id (:string) (required)

admin_contacts (:list) (required) - list of project admin emails.

can_edit_data_retention (:boolean) (required)

wiped_dataclip_viewer(assigns)

Attributes

id (:string) - Defaults to nil.

input_or_output (:atom) (required) - Must be one of :input, or :output.

project_id (:string) (required)

admin_contacts (:list) (required) - list of project admin emails.

can_edit_data_retention (:boolean) (required)

Slots

footer