LightningWeb.JobLive.FormComponent behaviour (Lightning v0.1.7)

Macro for Form Components that edit and show Jobs.

Callbacks

save( job_params, socket )

@callback save(
  job_params :: Phoenix.LiveView.unsigned_params(),
  socket :: Phoenix.LiveView.Socket.t()
) :: Phoenix.LiveView.Socket.t()

validate( job_params, socket )

@callback validate(
  job_params :: Phoenix.LiveView.unsigned_params(),
  socket :: Phoenix.LiveView.Socket.t()
) :: Phoenix.LiveView.Socket.t()

Functions

coerce_params_for_adaptor_list(job_params)

@spec coerce_params_for_adaptor_list(%{required(String.t()) => String.t()}) :: %{}

Coerce any changes to the "Adaptor" dropdown into a new selection on the Version dropdown.

get_adaptor_version_options(adaptor)

requires_cron_job?(changeset)

requires_upstream_job?(changeset)

save(params, socket)

update(assigns, socket)

validate(map, socket)